% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_content_type.R
\name{add_content_type}
\alias{add_content_type}
\title{declare a new content type}
\usage{
add_content_type(x, extension, type)
}
\arguments{
\item{x}{a docx object}

\item{extension}{extension string value (i.e. 'png')}

\item{type}{corresponding type (i.e. 'image/png')}
}
\description{
The function add a new content type associated to an
extension into a docx object.
}
\examples{
x <- docx()
x <- add_content_type(x, extension = "png", type = "image/png")
}
