% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_notes.R
\name{add_notes}
\alias{add_notes}
\alias{add_notes.character}
\title{add notes to a slide}
\usage{
add_notes(x, value, location, ...)

\method{add_notes}{character}(x, value, location, ...)
}
\arguments{
\item{x}{an rpptx object}
}
\description{
add speaker notes to a slide in a pptx presentation
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: add a character vector to a place holder in the notes on the
current slide, values will be added as paragraphs.
}}

\seealso{
\code{\link[=print.rpptx]{print.rpptx()}}, \code{\link[=read_pptx]{read_pptx()}}, \code{\link[=add_slide]{add_slide()}}

Other functions slide manipulation: 
\code{\link{add_slide}()},
\code{\link{move_slide}()},
\code{\link{on_slide}()},
\code{\link{remove_slide}()}
}
\concept{functions slide manipulation}
