% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ppt.R
\name{add_ts}
\alias{add_ts}
\title{Add a title slide}
\usage{
add_ts(wb, type = NULL, title = NULL, author = NULL, date = NULL)
}
\arguments{
\item{wb}{A \code{pptWorkbook}.}

\item{type}{The type of report (blue text).}

\item{title}{Title for the report (red text).}

\item{author}{The first line after the title, usually author.}

\item{date}{Last line on the title slide. E.g. year.}
}
\description{
Use \code{add_ts} to add a title slide to a \code{pptWorkbook}.
}
\author{
Kristian D. Olsen
}
\seealso{
\code{\link{to_ppt}} for more information.
}

