% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_content.R
\docType{data}
\name{any_content}
\alias{any_content}
\title{generic container}
\format{a dummy generic container}
\usage{
any_content
}
\section{Fields}{

\describe{
\item{\code{new(node)}}{create a container, if \code{node} is missing an empty container is returned.}

\item{\code{set_content(text)}}{set content, \code{node} is containing xml code in a single string.}

\item{\code{format()}}{get xml code}

\item{\code{print()}}{print object}
}}
\keyword{datasets}

