% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev-utils.R
\name{as_base64}
\alias{as_base64}
\title{Encode Character Vector to Base64}
\usage{
as_base64(x)
}
\arguments{
\item{x}{A character vector. NA values are preserved.}
}
\value{
A character vector of Base64-encoded strings, same length as \code{x}.
}
\description{
Encodes one or more elements of a character vector into Base64 format.
}
\examples{
as_base64(letters)
as_base64(c("hello", NA, "world"))
}
\keyword{internal}
