% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_list.R
\name{block_list}
\alias{block_list}
\title{list of blocks}
\usage{
block_list(...)
}
\arguments{
\item{...}{a list of objects of class \code{fpar} or
\code{flextable}.}
}
\description{
a list of blocks can be used to gather
several blocks (paragraphs or tables) into a single
object. The function is to be used when adding
footnotes for example.
}
\examples{
img.file <- file.path( R.home("doc"), "html", "logo.jpg" )
bl <- block_list(
  fpar(ftext("hello world", shortcuts$fp_bold())),
  fpar(
    ftext("hello", shortcuts$fp_bold()),
    stext(" world", "strong"),
    external_img(src = img.file, height = 1.06, width = 1.39)
  )
)
}
