% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_block_objects.R
\name{block_toc}
\alias{block_toc}
\title{table of content}
\usage{
block_toc(level = 3, style = NULL, separator = ";")
}
\arguments{
\item{level}{max title level of the table}

\item{style}{optional. style in the document that will be used to build entries of the TOC.}

\item{separator}{optional. Some configurations need "," (i.e. from Canada) separator instead of ";"}
}
\description{
Create a representation of a table of content.
}
\examples{
block_toc(level = 2)
block_toc(style = "Table title")
}
\seealso{
Other block functions for reporting: 
\code{\link{block_caption}()},
\code{\link{block_list}()},
\code{\link{block_section}()},
\code{\link{block_table}()},
\code{\link{fpar}()},
\code{\link{plot_instr}()},
\code{\link{unordered_list}()}
}
\concept{block functions for reporting}
