% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_add.R
\name{body_add_break}
\alias{body_add_break}
\title{add page break}
\usage{
body_add_break(x, pos = "after")
}
\arguments{
\item{x}{an rdocx object}

\item{pos}{where to add the new element relative to the cursor,
one of "after", "before", "on".}
}
\description{
add a page break into an rdocx object
}
\examples{
library(magrittr)
doc <- read_docx() \%>\% body_add_break()
print(doc, target = tempfile(fileext = ".docx"))
}
