% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_add.R
\name{body_add_plot}
\alias{body_add_plot}
\title{add plot}
\usage{
body_add_plot(x, value, width = 6, height = 5, res = 300, style = NULL, ...)
}
\arguments{
\item{x}{an rdocx object}

\item{value}{plotting instructions (an expression)}

\item{width}{height in inches}

\item{height}{height in inches}

\item{res}{resolution of the png image in ppi}

\item{style}{paragraph style}

\item{...}{Arguments to be passed to png function.}
}
\description{
add a plot as a png image into an rdocx object
}
\examples{
doc <- read_docx()

if( capabilities(what = "png") )
  doc <- body_add_plot(doc, value = barplot(1:6), style = "centered" )

print(doc, target = tempfile(fileext = ".docx") )
}
