% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_add.R
\name{body_add_xml_multi}
\alias{body_add_xml_multi}
\title{Add several xml strings as document elements}
\usage{
body_add_xml_multi(x, strv, pos = c("after", "before", "on"), pagebreak = TRUE)
}
\arguments{
\item{x}{an rdocx object}

\item{strv}{a vevtor of wml strings}

\item{pos}{where to add the new element relative to the cursor,
one of "after", "before", "on".}

\item{pagebreak}{insert page breaks between WML elements provided}
}
\description{
This method applies the same actions as \code{body_add_xml()} but takes in
a vector of XML elements.
}
\keyword{internal}
