% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_add.R
\name{body_bookmark}
\alias{body_bookmark}
\title{add bookmark}
\usage{
body_bookmark(x, id)
}
\arguments{
\item{x}{an rdocx object}

\item{id}{bookmark name}
}
\description{
Add a bookmark at the cursor location. The bookmark
is added on the first run of text in the current paragraph.
}
\examples{

# cursor_bookmark ----
library(magrittr)

doc <- read_docx() \%>\%
  body_add_par("centered text", style = "centered") \%>\%
  body_bookmark("text_to_replace")
}
