% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev-docx_add.R
\name{body_comment}
\alias{body_comment}
\title{Add comment in a 'Word' document}
\usage{
body_comment(x, value)
}
\arguments{
\item{x}{an rdocx object}

\item{id}{bookmark name}
}
\description{
Add a comment at the cursor location. The comment
is added on the first run of text in the current paragraph.
}
\examples{

# cursor_bookmark ----

doc <- read_docx()
doc <- body_add_par(doc, "centered text", style = "centered")
doc <- body_comment(doc, )
}
