% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_add.R
\name{body_replace_all_text}
\alias{body_replace_all_text}
\title{replace text anywhere in the document, or at a cursor}
\usage{
body_replace_all_text(x, oldValue, newValue, onlyAtCursor = FALSE,
  mergeRuns = FALSE)
}
\arguments{
\item{x}{a docx device}

\item{oldValue}{a string to replace}

\item{newValue}{the string to replace it with}

\item{onlyAtCursor}{only search-and-replace at the current cursor}

\item{mergeRuns}{merge runs before trying search-and-replace
library(magrittr)
doc <- read_docx() %>%
  body_replace_all_text("my_placeholder_text", "new text")}
}
\description{
replace all occurrences of oldValue with newValue
}
