% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_replace.R
\name{body_replace_text_at_bkm}
\alias{body_replace_text_at_bkm}
\alias{body_replace_at}
\alias{body_replace_img_at_bkm}
\alias{headers_replace_text_at_bkm}
\alias{headers_replace_img_at_bkm}
\alias{footers_replace_text_at_bkm}
\alias{footers_replace_img_at_bkm}
\title{replace text at a bookmark location}
\usage{
body_replace_text_at_bkm(x, bookmark, value)

body_replace_at(x, bookmark, value)

body_replace_img_at_bkm(x, bookmark, value)

headers_replace_text_at_bkm(x, bookmark, value)

headers_replace_img_at_bkm(x, bookmark, value)

footers_replace_text_at_bkm(x, bookmark, value)

footers_replace_img_at_bkm(x, bookmark, value)
}
\arguments{
\item{x}{a docx device}

\item{bookmark}{bookmark id}

\item{value}{the replacement string, of type character}
}
\description{
replace text content enclosed in a bookmark
with different text. A bookmark will be considered as valid if enclosing words
within a paragraph; i.e., a bookmark along two or more paragraphs is invalid,
a bookmark set on a whole paragraph is also invalid, but bookmarking few words
inside a paragraph is valid.
}
\examples{
library(magrittr)
doc <- read_docx() \%>\%
  body_add_par("centered text", style = "centered") \%>\%
  slip_in_text(". How are you", style = "strong") \%>\%
  body_bookmark("text_to_replace") \%>\%
  body_replace_text_at_bkm("text_to_replace", "not left aligned")


# demo usage of bookmark and images ----
template <- system.file(package = "officer", "doc_examples/example.docx")

img.file <- file.path( R.home("doc"), "html", "logo.jpg" )

doc <- read_docx(path = template)
doc <- headers_replace_img_at_bkm(x = doc, bookmark = "bmk_header",
                                  value = external_img(src = img.file, width = .53, height = .7))
doc <- footers_replace_img_at_bkm(x = doc, bookmark = "bmk_footer",
                                  value = external_img(src = img.file, width = .53, height = .7))
print(doc, target = tempfile(fileext = ".docx"))

}
