% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_add.R
\name{bookmark}
\alias{bookmark}
\title{docx bookmark xml}
\usage{
bookmark(id, str)
}
\arguments{
\item{id}{bookmark name}

\item{str}{a scalar character representing an xml \emph{run}.}
}
\description{
Create a string representation of a bookmark.
}
\examples{
bookmark("<r><t>hi</t></r>")
}
