% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx_slide_manip.R
\name{clean_unused_media}
\alias{clean_unused_media}
\title{Clean unused media files}
\usage{
clean_unused_media(x)
}
\arguments{
\item{x}{an rpptx object}
}
\value{
an rpptx object (invisibly)
}
\description{
Remove media files that are no longer referenced by any slides or components
in the PowerPoint presentation. This helps reduce file size by removing orphaned images.
}
\examples{
# Create a presentation and add some content
my_pres <- read_pptx()
my_pres <- add_slide(my_pres, "Title and Content")

# After removing slides or content, clean up unused media
my_pres <- clean_unused_media(my_pres)
}
\seealso{
\code{\link[=read_pptx]{read_pptx()}}, \code{\link[=remove_slide]{remove_slide()}}, \code{\link[=remove_slides_bulk]{remove_slides_bulk()}}

Other functions to manipulate slides:
\code{\link{add_slide}()},
\code{\link{move_slide}()},
\code{\link{on_slide}()},
\code{\link{remove_slide}()},
\code{\link{remove_slides_bulk}()},
\code{\link{set_notes}()}
}
\concept{slide_manipulation}
