% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_cursor.R
\name{cursor_begin}
\alias{cursor_begin}
\alias{cursor_end}
\alias{cursor_reach}
\alias{cursor_forward}
\alias{cursor_backward}
\title{set cursor in a docx object}
\usage{
cursor_begin(x)

cursor_end(x)

cursor_reach(x, keyword)

cursor_forward(x)

cursor_backward(x)
}
\arguments{
\item{x}{a docx device}

\item{keyword}{keyword to look for}
}
\description{
a set of functions is available to manipulate
the position of a virtual cursor. This cursor will be used when
inserting, deleting or updating elements in the document.
}
\section{cursor_begin}{

Set the cursor at the beginning of the document, on the first element
of the document (usually a paragraph or a table).
}

\section{cursor_end}{

Set the cursor at the end of the document, on the last element
of the document.
}

\section{cursor_reach}{

Set the cursor on the first element of the document
that contains text specified in argument \code{keyword}.
}

\section{cursor_forward}{

Move the cursor forward, it increments the cursor in the document.
}

\section{cursor_backward}{

Move the cursor backward, it decrements the cursor in the document.
}

