% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_add.R
\name{docx_add_img}
\alias{docx_add_img}
\title{add an image}
\usage{
docx_add_img(x, src, style = "Normal", width, height, pos = "after")
}
\arguments{
\item{x}{a docx device}

\item{src}{image filename}

\item{style}{paragraph style}

\item{width}{height in inches}

\item{height}{height in inches}

\item{pos}{where to add the new element relative to the cursor,
one of "after", "before", "on".}
}
\description{
add an image into a docx object
}

