% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_add.R
\name{docx_add_pbreak}
\alias{docx_add_pbreak}
\title{add a page break}
\usage{
docx_add_pbreak(x, pos = "after")
}
\arguments{
\item{x}{a docx object}

\item{pos}{where to add the new element relative to the cursor,
one of "after", "before", "on".}
}
\description{
add a page break into a docx object
}
