% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_add.R
\name{docx_add_toc}
\alias{docx_add_toc}
\title{add a table of content}
\usage{
docx_add_toc(x, level = 3, pos = "after", style = NULL, separator = ";")
}
\arguments{
\item{x}{a docx object}

\item{level}{max title level of the table}

\item{pos}{where to add the new element relative to the cursor,
one of "after", "before", "on".}

\item{style}{optional. style in the document that will be used to build entries of the TOC.}

\item{separator}{optional. Some configurations need "," (i.e. from Canada) separator instead of ";"}
}
\description{
add a table of content into a docx object
}

