% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_add_chunk.R
\name{docx_append_run}
\alias{docx_append_run}
\title{append text}
\usage{
docx_append_run(x, str, style = "Normal", pos = "after")
}
\arguments{
\item{x}{a docx device}

\item{str}{text}

\item{style}{text style}

\item{pos}{where to add the new element relative to the cursor,
"after" or "before".}
}
\description{
append text into a paragraph of a docx object
}
\examples{
library(magrittr)
docx() \%>\%
  docx_add_par("Hello ", style = "Normal") \%>\%
  docx_append_run("world", style = "strong") \%>\%
  docx_append_run("Message is", style = "strong", pos = "before") \%>\%
  print(target = "append_run.docx")
}

