% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_add_chunk.R
\name{docx_append_seqfield}
\alias{docx_append_seqfield}
\title{append seq field}
\usage{
docx_append_seqfield(x, str, style = "Normal", pos = "after")
}
\arguments{
\item{x}{a docx device}

\item{str}{seq field value}

\item{style}{text style}

\item{pos}{where to add the new element relative to the cursor,
"after" or "before".}
}
\description{
append seq field into a paragraph of a docx object
}
\examples{
library(magrittr)
docx() \%>\%
  docx_add_par("Time is: ", style = "Normal") \%>\%
  docx_append_seqfield(
    str = "TIME \\u005C@ \\"HH:mm:ss\\" \\u005C* MERGEFORMAT",
    style = 'strong') \%>\%

  docx_add_par(" - This is a figure title", style = "centered") \%>\%
  docx_append_seqfield(str = "SEQ Figure \\u005C* roman",
    style = 'Default Paragraph Font', pos = "before") \%>\%
  docx_append_run("Figure: ", style = "strong", pos = "before") \%>\%

  docx_add_par(" - This is another figure title", style = "centered") \%>\%
  docx_append_seqfield(str = "SEQ Figure \\u005C* roman",
    style = 'strong', pos = "before")  \%>\%
  docx_append_run("Figure: ", style = "strong", pos = "before") \%>\%
  docx_add_par("This is a symbol: ", style = "Normal") \%>\%
  docx_append_seqfield(str = "SYMBOL 100 \\u005Cf Wingdings",
    style = 'strong') \%>\%
  print(target = "seqfield.docx")
}
