% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_docx.R
\name{docx_body_xml}
\alias{docx_body_xml}
\title{body xml document}
\usage{
docx_body_xml(x)
}
\arguments{
\item{x}{an rdocx object}
}
\description{
Get the body document as xml. This function
is not to be used by end users, it has been implemented
to allow other packages to work with officer.
}
\examples{
doc <- read_docx()
docx_body_xml(doc)
}
\keyword{internal}
