% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_docx.R
\name{docx_bookmarks}
\alias{docx_bookmarks}
\title{List Word bookmarks}
\usage{
docx_bookmarks(x)
}
\arguments{
\item{x}{an \code{rdocx} object}
}
\description{
List bookmarks id that can be found in a
'Word' document.
}
\examples{
library(officer)

doc_1 <- read_docx()
doc_1 <- body_add_par(doc_1, "centered text", style = "centered")
doc_1 <- body_bookmark(doc_1, "text_to_replace_1")
doc_1 <- body_add_par(doc_1, "centered text", style = "centered")
doc_1 <- body_bookmark(doc_1, "text_to_replace_2")

docx_bookmarks(doc_1)

docx_bookmarks(read_docx())
}
\seealso{
Other functions for Word document informations: 
\code{\link{doc_properties}()},
\code{\link{docx_dim}()},
\code{\link{length.rdocx}()},
\code{\link{set_doc_properties}()},
\code{\link{styles_info}()}
}
\concept{functions for Word document informations}
