% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_docx.R
\name{docx_current_block_xml}
\alias{docx_current_block_xml}
\title{xml element on which cursor is}
\usage{
docx_current_block_xml(x)
}
\arguments{
\item{x}{an rdocx object}
}
\description{
Get the current block element as xml. This function
is not to be used by end users, it has been implemented
to allow other packages to work with officer. If the
document is empty, this block will be set to NULL.
}
\examples{
doc <- read_docx()
docx_current_block_xml(doc)
}
\keyword{internal}
