% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_docx.R
\name{docx_dim}
\alias{docx_dim}
\title{Word page layout}
\usage{
docx_dim(x)
}
\arguments{
\item{x}{a \code{rdocx} object}
}
\description{
get page width, page height and margins (in inches). The return values
are those corresponding to the section where the cursor is.
}
\examples{
docx_dim(read_docx())
}
