% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_img.R
\name{docx_reference_img}
\alias{docx_reference_img}
\title{add images into an rdocx object}
\usage{
docx_reference_img(x, src)
}
\arguments{
\item{x}{an rdocx object}

\item{src}{a vector of character containing image filenames.}
}
\description{
reference images into a Word document.
This function is to be used with \code{\link{wml_link_images}}.

Images need to be referenced into the
Word document, this will generate unique
identifiers that need to be known to
link these images with their corresponding xml code (wml).
}
