% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_add.R
\name{docx_show_chunk}
\alias{docx_show_chunk}
\title{show underlying text tag structure}
\usage{
docx_show_chunk(x)
}
\arguments{
\item{x}{a docx device}
}
\description{
show the structure of text tags at the current cursor
}
\examples{
library(magrittr)
doc <- read_docx() \%>\%
  docx_show_chunk(doc)
}
