% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_run_objects.R
\name{external_img}
\alias{external_img}
\title{External image}
\usage{
external_img(src, svg_src = NULL, width = 0.5, height = 0.2, alt = "")
}
\arguments{
\item{src}{image file path}

\item{svg_src}{used internally.  See note.}

\item{width}{height in inches.}

\item{height}{height in inches}

\item{alt}{alternative text for images}
}
\description{
Wraps an image in an object that can then be embedded
in a PowerPoint slide or within a Word paragraph.

The image is added as a shape in PowerPoint (it is not possible to mix text and
images in a PowerPoint form). With a Word document, the image will be
added inside a paragraph.
}
\note{
For internal use (meaning maintaining the package):
\code{svg_src} is the svg image file path for adding svg to Word/PPT.
must still specify \code{src} for non-svg (png) image path as fallback for older versions of Word/PPT.
User-facing code can just specify the svg file to \code{src}.
}
\section{usage}{

You can use this function in conjunction with \link{fpar} to create paragraphs
consisting of differently formatted text parts. You can also use this
function as an \emph{r chunk} in an R Markdown document made with package
officedown.
}

\examples{
# wrap r logo with external_img ----
srcfile <- file.path( R.home("doc"), "html", "logo.jpg" )
extimg <- external_img(src = srcfile, height = 1.06/2,
                       width = 1.39/2)

# pptx example ----
doc <- read_pptx()
doc <- add_slide(doc)
doc <- ph_with(x = doc, value = extimg,
               location = ph_location_type(type = "body"),
               use_loc_size = FALSE )
print(doc, target = tempfile(fileext = ".pptx"))

fp_t <- fp_text(font.size = 20, color = "red")
an_fpar <- fpar(extimg, ftext(" is cool!", fp_t))

# docx example ----
x <- read_docx()
x <- body_add(x, an_fpar)
print(x, target = tempfile(fileext = ".docx"))

# svg example docx ----
svgfile <- file.path( R.home("doc"), "html", "Rlogo.svg" )
ext_svg <- external_img(src = svgfile, height = 1.06/2, width = 1.39/2)
x <- read_docx()
x <- body_add(x, ext_svg)
print(x, target = tempfile(fileext = ".docx"))
# svg example pptx ----
x <- read_pptx()
x <- add_slide(x)
x <- ph_with(x, ext_svg, location = ph_location_type(), alt_text = alt_text, height = 1.06, width = 1.39, use_loc_size=FALSE) # ph_location_type default is body
ppt <- tempfile(fileext = ".pptx")
print(x, target = 'temp.pptx')



}
\seealso{
\link{ph_with}, \link{body_add}, \link{fpar}

Other run functions for reporting: 
\code{\link{ftext}()},
\code{\link{hyperlink_ftext}()},
\code{\link{run_autonum}()},
\code{\link{run_bookmark}()},
\code{\link{run_columnbreak}()},
\code{\link{run_footnoteref}()},
\code{\link{run_footnote}()},
\code{\link{run_linebreak}()},
\code{\link{run_pagebreak}()},
\code{\link{run_reference}()},
\code{\link{run_word_field}()}
}
\concept{run functions for reporting}
