% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_sign.R
\name{fp_sign}
\alias{fp_sign}
\title{object unique signature}
\usage{
fp_sign(x)
}
\arguments{
\item{x}{a formatting set of properties}
}
\description{
Get unique signature for a formatting properties
object.
}
\examples{
fp_sign( fp_text(color="orange") )
}
