% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpar.R
\name{fpar}
\alias{fpar}
\alias{update.fpar}
\alias{dim.fpar}
\alias{format.fpar}
\alias{print.fpar}
\title{formatted paragraph}
\usage{
fpar(...)

\method{update}{fpar}(object, fp_p = NULL, fp_t = NULL, ...)

\method{dim}{fpar}(x)

\method{format}{fpar}(x, type = "pml", ...)

\method{print}{fpar}(x, ...)
}
\arguments{
\item{...}{unused}

\item{fp_p}{paragraph formatting properties}

\item{fp_t}{default text formatting properties}

\item{x, object}{fpar object}

\item{type}{a string value ("pml", "wml" or "html").}
}
\description{
formatted paragraph
}
\examples{
fpar(ftext("hello", fp_bold()))
}

