% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled.R
\name{from_labelled}
\alias{from_labelled}
\alias{from_labelled.data.frame}
\alias{from_labelled.data.table}
\title{Convert from labelled data}
\usage{
from_labelled(df)

\method{from_labelled}{data.frame}(df)

\method{from_labelled}{data.table}(df)
}
\arguments{
\item{df}{A data.frame as returned from \code{read_data} or
\code{\link[haven]{read_sav}}.}
}
\description{
When reading SPSS files with \code{\link{read_data}}, the output contains vectors
of class \code{labelled}. This function extracts the labels and attaches them
as an attribute (by the same name) to the data, and converts any \code{labelled}
variables into factors.
}
\note{
\code{\link[data.table]{data.table}} input returns a copy of the \code{data.table}.
}
\examples{
# TODO
}
\author{
Kristian D. Olsen
}

