% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx_layout_helper.R
\name{get_layout}
\alias{get_layout}
\title{Layout selection helper}
\usage{
get_layout(x, layout, master = NULL)
}
\arguments{
\item{x}{An \code{rpptx} object.}

\item{layout}{Layout name or index, see \code{\link[=layout_summary]{layout_summary()}}.}

\item{master}{Name of master. Only required if layout name is not unique across masters.}
}
\value{
A list with the entries index, layout_name, layout_file, master_name, master_file
}
\description{
Select a layout by name or index. The master name is inferred and only required
for disambiguation in case the layout name is not unique across masters.
}
\examples{
x <- read_pptx()
get_layout(x, "Title Slide")

}
\keyword{internal}
