% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_pptx.R
\name{get_reference_pptx}
\alias{get_reference_pptx}
\title{Get the document being used as a template}
\usage{
get_reference_pptx()
}
\description{
Get filename of the PowerPoint document
being used as a template in an R Markdown document
render as PowerPoint presentation. It requires
packages rmarkdown >= 1.10.14 and knitr.
}
\author{
Noam Ross
}
