% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_utils.R
\name{get_reference_value}
\alias{get_reference_value}
\title{Get the document being used as a template}
\usage{
get_reference_value(format = NULL)
}
\arguments{
\item{format}{document format, one of 'pptx', 'docx' or 'html'}
}
\value{
a name file
}
\description{
Get filename of the document
being used as a template in an R Markdown document
rendered as HTML, PowerPoint presentation or Word document. It requires
packages rmarkdown >= 1.10.14 and knitr.
}
\seealso{
Other functions for officer extensions: 
\code{\link{docx_reference_img}()},
\code{\link{fortify_location}()},
\code{\link{opts_current_table}()},
\code{\link{to_pml}()},
\code{\link{to_wml}()},
\code{\link{wml_link_images}()}
}
\concept{functions for officer extensions}
\keyword{internal}
