% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev-utils.R
\name{image_to_base64}
\alias{image_to_base64}
\title{Images to base64}
\usage{
image_to_base64(filepaths)
}
\arguments{
\item{filepaths}{file names.}
}
\description{
encodes images into base64 strings.
}
\examples{
rlogo <- file.path( R.home("doc"), "html", "logo.jpg")
image_to_base64(rlogo)
}
\keyword{internal}
