% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_utils_funs.R
\name{length.rdocx}
\alias{length.rdocx}
\title{Number of blocks inside an rdocx object}
\usage{
\method{length}{rdocx}(x)
}
\arguments{
\item{x}{an rdocx object}
}
\description{
Return the number of blocks inside an rdocx object.
This number also include the default section definition of a
Word document - default Word section is an uninvisible element.
}
\examples{
# how many elements are there in an new document produced
# with the default template.
length( read_docx() )
}
\seealso{
Other functions for Word document informations: 
\code{\link{doc_properties}()},
\code{\link{docx_bookmarks}()},
\code{\link{docx_dim}()},
\code{\link{set_doc_properties}()},
\code{\link{styles_info}()}
}
\concept{functions for Word document informations}
