% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_block_objects.R
\name{list_item}
\alias{list_item}
\title{Create a list item}
\usage{
list_item(x, level = 1L)
}
\arguments{
\item{x}{an \code{\link[=fpar]{fpar()}} object or a character string
(automatically converted to \code{\link[=fpar]{fpar()}})}

\item{level}{hierarchy level, integer starting at 1}
}
\description{
Wrap an \code{\link[=fpar]{fpar()}} with a hierarchy level for use
inside \code{\link[=block_list_items]{block_list_items()}}.
}
\seealso{
\code{\link[=block_list_items]{block_list_items()}}

Other block functions for reporting: 
\code{\link{block_caption}()},
\code{\link{block_gg}()},
\code{\link{block_list}()},
\code{\link{block_list_items}()},
\code{\link{block_pour_docx}()},
\code{\link{block_section}()},
\code{\link{block_table}()},
\code{\link{block_toc}()},
\code{\link{fpar}()},
\code{\link{plot_instr}()},
\code{\link{unordered_list}()}
}
\concept{block functions for reporting}
