% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_notes.R
\name{notes_location_label}
\alias{notes_location_label}
\title{Location of a named placeholder for notes}
\usage{
notes_location_label(ph_label, ...)
}
\arguments{
\item{ph_label}{placeholder label of the used notes master}

\item{...}{unused arguments}
}
\description{
The function will use the label of a placeholder
to find the corresponding location in the slide notes.
}
