% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_notes.R
\name{notes_location_type}
\alias{notes_location_type}
\title{Location of a placeholder for notes}
\usage{
notes_location_type(type = "body", ...)
}
\arguments{
\item{type}{placeholder label of the used notes master}

\item{...}{unused arguments}
}
\description{
The function will use the type name of the placeholder (e.g. body, hdr),
to find the corresponding location.
}
