% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_add.R
\name{numbered_caption}
\alias{numbered_caption}
\alias{format.numbered_caption}
\title{numbered caption}
\usage{
numbered_caption(
  label,
  style,
  id,
  seq_id = "table",
  pre_label = "TABLE ",
  post_label = ": "
)

\method{format}{numbered_caption}(x, type = "docx", base_document = NULL, ...)
}
\arguments{
\item{label}{a scalar character representing label to display}

\item{style}{paragraph style identifier}

\item{id}{cross reference identifier}

\item{type}{output format, one of docx, html.}

\item{...}{unused}
}
\description{
Create a string representation of a numbered
caption that can be used for cross reference.
}
