% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{officer-defunct}
\alias{officer-defunct}
\alias{docx_reference_img}
\title{Defunct Functions in Package officer}
\usage{
docx_reference_img(...)
}
\arguments{
\item{...}{unused arguments}
}
\description{
Defunct Functions in Package officer
}
\details{
\code{docx_reference_img()} is not usefull anymore.
}
