% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev-utils.R
\name{officer_url_encode}
\alias{officer_url_encode}
\title{officer url encoder}
\usage{
officer_url_encode(x)
}
\arguments{
\item{x}{a character vector of URL}
}
\description{
encode url so that it can be easily
decoded when 'officer' write a file to the disk.
}
\examples{
officer_url_encode("https://cran.r-project.org/")
}
\keyword{internal}
