% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{open_office_file}
\alias{open_office_file}
\title{Opens a .pptx or .docx file locally}
\usage{
open_office_file(path)
}
\arguments{
\item{path}{File path.}
}
\description{
Opening a \code{.pptx} or \code{.docx} file locally requires a compatible application
(e.g., MS Office, LibreOffice) to be installed. \emph{Disclaimer}: may not work
properly on every system.
}
\examples{
x <- read_pptx()
x <- add_slide(x, "Title Slide", ctrTitle = "My Title")
file <- print(x, tempfile(fileext = ".pptx"))
\dontrun{
open_office_file(file)}
}
