% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_utils.R
\name{opts_current_table}
\alias{opts_current_table}
\title{Get table options in a 'knitr' context}
\usage{
opts_current_table()
}
\value{
a list with following elements:
\itemize{
\item cap.style (default: NULL)
\item cap.pre (default: "Table ")
\item cap.sep (default: ":")
\item id (default: NULL)
\item cap (default: NULL)
\item style (default: NULL)
\item tab.lp (default: "tab:")
\item table_layout (default: "autofit")
\item table_width (default: 1)
\item first_row (default: TRUE)
\item first_column (default: FALSE)
\item last_row (default: FALSE)
\item last_column (default: FALSE)
\item no_hband (default: TRUE)
\item no_vband (default: TRUE)
}
}
\description{
Get options for table rendering.

It should not be used by the end user.
The function is a utility to facilitate the retrieval of table
options supported by the 'flextable', 'officedown' and of
course 'officer' packages.
}
\seealso{
Other functions for officer extensions: 
\code{\link{docx_reference_img}()},
\code{\link{fortify_location}()},
\code{\link{get_reference_value}()},
\code{\link{to_pml}()},
\code{\link{to_wml}()},
\code{\link{wml_link_images}()}
}
\concept{functions for officer extensions}
\keyword{internal}
