% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_folder.R
\name{pack_folder}
\alias{pack_folder}
\title{compress a folder}
\usage{
pack_folder(folder, target)
}
\arguments{
\item{folder}{folder to compress}

\item{target}{path of the archive to create}
}
\description{
compress a folder to a target file. The
function returns the complete path to target file.
}
\details{
The function is using \link[utils]{zip}, it needs a zip program.
}
