% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_run_objects.R
\name{page_size}
\alias{page_size}
\title{Page size object}
\usage{
page_size(
  width = 11906/1440,
  height = 16838/1440,
  orient = "portrait",
  unit = "in"
)
}
\arguments{
\item{width, height}{page width, page height, default to A4 format If NULL
the value will be ignored and Word will use the default value.}

\item{orient}{page orientation, either 'landscape', either 'portrait'.}

\item{unit}{unit for width and height, one of "in", "cm", "mm".}
}
\description{
The function creates a representation of the dimensions of a page. The
dimensions are defined by length, width and orientation. If the orientation is
in landscape mode then the length becomes the width and the width becomes the length.
}
\examples{
page_size(orient = "landscape")
}
\seealso{
Other functions for section definition: 
\code{\link{page_mar}()},
\code{\link{prop_section}()},
\code{\link{section_columns}()}
}
\concept{functions for section definition}
