% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\docType{data}
\name{paragraph}
\alias{paragraph}
\title{paragraph container}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
paragraph
}
\description{
a paragraph container
}
\keyword{datasets}

