% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_ph_add.R
\name{ph_add_text2}
\alias{ph_add_text2}
\title{append text v2}
\usage{
ph_add_text2(x, str, type = NULL, id_chr = NULL, style = fp_text(font.size
  = 0), pos = "after", slide_index = NULL)
}
\arguments{
\item{x}{a pptx device}

\item{str}{text to add}

\item{type}{placeholder type}

\item{id_chr}{placeholder id (a string). This is to be used when a placeholder type
is not unique in the current slide, e.g. two placeholders with type 'body'.
Values can be read from \code{\link{slide_summary}}.}

\item{style}{text style, a \code{\link{fp_text}} object}

\item{pos}{where to add the new element relative to the cursor,
"after" or "before".}

\item{slide_index}{slide index to reach}
}
\description{
append text in a placeholder, but here link to slide is possible instead of hyperlink
}
\examples{
library(magrittr)
fileout <- tempfile(fileext = ".pptx")
doc <- read_pptx() \%>\%
  add_slide(layout = "Title and Content", master = "Office Theme") \%>\%
  ph_with_text(type = "title", str = "Un titre 1") \%>\%
  add_slide(layout = "Title and Content", master = "Office Theme") \%>\%
  ph_with_text(type = "title", str = "Un titre 2") \%>\%
  on_slide(1) \%>\%
  ph_empty(type = "body") \%>\%
  ph_add_par(type = "body", level = 2) \%>\%
  ph_add_text2(str = "Jump here to slide 2!", type = "body", slide_index=2)
print(doc, target = fileout)
}
