% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_ph_manipulate.R
\name{ph_hyperlink}
\alias{ph_hyperlink}
\title{hyperlink a placeholder}
\usage{
ph_hyperlink(x, type = "body", id = 1, id_chr = NULL,
  ph_label = NULL, href)
}
\arguments{
\item{x}{a pptx device}

\item{type}{placeholder type}

\item{id}{placeholder index for a duplicated type(integer). This is to be used when a placeholder
type is not unique in the layout of the current slide, e.g. two placeholders with type 'body'. To
add onto the first, use \code{id = 1} and \code{id = 2} for the second one.
Values can be read from \code{\link{slide_summary}}.}

\item{id_chr}{deprecated.}

\item{ph_label}{label associated to the placeholder. Use column
\code{ph_label} of result returned by \code{\link{slide_summary}}.}

\item{href}{hyperlink (do not forget http or https prefix)}
}
\description{
add hyperlink to a placeholder in the current slide.
}
\examples{
fileout <- tempfile(fileext = ".pptx")
doc <- read_pptx()
doc <- add_slide(doc, layout = "Title and Content", master = "Office Theme")
doc <- ph_with_text(x = doc, type = "title", str = "Un titre 1")
slide_summary(doc) # read column ph_label here
doc <- ph_hyperlink(x = doc, ph_label = "Title 1",
  href = "https://cran.r-project.org")

print(doc, target = fileout )
}
