% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_location.R
\name{ph_location}
\alias{ph_location}
\title{create a location for a placeholder}
\usage{
ph_location(left, top, width, height, label)
}
\arguments{
\item{x}{an rpptx object}

\item{ph_label}{a label for the placeholder.}
}
\description{
The function will return a list that complies with
expected format for argument \code{location} of functions \code{ph_with_*}.
}
\details{
The location of the bounding box associated to a placeholder
within a presentation slide is specified with the left top coordinate,
the width and the height. These are defined in inches:

\describe{
  \item{offx}{left coordinate of the bounding box}
  \item{offy}{top coordinate of the bounding box}
  \item{width}{width of the bounding box}
  \item{height}{height of the bounding box}
}

In addition to these attributes, there is also an attribute \code{ph_label}
that can be associated with the shape (shapes, text boxes, images and other objects
will be identified with that label in the Selection Pane of PowerPoint).

This result can then be used to position new elements in the slides with
functions \code{ph_with_*}.
}
\examples{
x <- read_pptx()
ph_location_fullsize(x)
}
\seealso{
Other functions for placeholder location: \code{\link{ph_location_fullsize}},
  \code{\link{ph_location_label}},
  \code{\link{ph_location_left}},
  \code{\link{ph_location_right}},
  \code{\link{ph_location_type}}
}
\concept{functions for placeholder location}
