% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_ph_with_methods.R
\name{ph_with}
\alias{ph_with}
\alias{ph_with.character}
\alias{ph_with.data.frame}
\alias{ph_with.gg}
\title{add object into a new shape}
\usage{
ph_with(x, value, location, ...)

\method{ph_with}{character}(x, value, location, ...)

\method{ph_with}{data.frame}(x, value, location, header = TRUE,
  first_row = TRUE, first_column = FALSE, last_row = FALSE,
  last_column = FALSE, ...)

\method{ph_with}{gg}(x, value, location, ...)
}
\arguments{
\item{x}{a pptx device}

\item{value}{object to add as a new shape.}

\item{location}{a placeholder location object. This is a convenient
argument that can replace usage of arguments \code{type} and \code{index}.
See section \code{see also}.}

\item{...}{Arguments to be passed to methods}

\item{header}{display header if TRUE}

\item{first_row, last_row, first_column, last_column}{logical for PowerPoint table options}
}
\description{
add object into a new shape in a slide.
}
\section{with character}{

When value is a character vector, each value will be
added as a paragraph.
}

\section{with data.frame}{

When value is a data.frame, a simple table
is added, use package \code{flextable} instead
for more advanced formattings.
}

\section{with a ggplot object}{

When value is a ggplot object, a raster plot
is produced and added, use package \code{rvg}
instead for more advanced graphical features.
}

\examples{
library(magrittr)

fileout <- tempfile(fileext = ".pptx")
doc <- read_pptx()
doc <- add_slide(doc, layout = "Title and Content",
  master = "Office Theme")

if( require("ggplot2") ){
  gg_plot <- ggplot(data = iris ) +
    geom_point(mapping = aes(Sepal.Length, Petal.Length),
      size = 3) +
    theme_minimal()
  doc <- ph_with(x = doc, value = gg_plot,
                 location = ph_location_fullsize(doc) )
}

doc <- ph_with(x = doc, value = c("Un titre", "Deux titre"),
               location = ph_location_left(doc) )
doc <- ph_with(x = doc, value = iris[1:4, 3:5],
               location = ph_location_right(doc) )

print(doc, target = fileout )
}
\seealso{
\link{ph_location_type}, \link{ph_location}, \link{ph_location_label},
\link{ph_location_left}, \link{ph_location_right}, \link{ph_location_fullsize}
}
