% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_ph_with_methods.R
\name{ph_with_fpars_at}
\alias{ph_with_fpars_at}
\title{add multiple formated paragraphs}
\usage{
ph_with_fpars_at(x, fpars = list(), fp_pars = list(), left, top, width,
  height, bg = "transparent", rot = 0, template_type = NULL,
  template_index = 1)
}
\arguments{
\item{x}{rpptx object}

\item{fpars}{list of \code{\link{fpar}} objects}

\item{fp_pars}{list of \code{\link{fp_par}} objects. The list can contain
NULL to keep defaults.}

\item{left, top}{location of the new shape on the slide}

\item{width, height}{shape size in inches}

\item{bg}{background color}

\item{rot}{rotation angle}

\item{template_type}{placeholder template type. If used, the new shape will
inherit the style from the placeholder template. If not used, no text
property is defined and for example text lists will not be indented.}

\item{template_index}{placeholder template index (integer). To be used when a placeholder
template type is not unique in the current slide, e.g. two placeholders with
type 'body'.}
}
\description{
add several formated paragraphs in a new shape in the current slide.
}
