% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{ph_with_img_at}
\alias{ph_with_img_at}
\title{add image}
\usage{
ph_with_img_at(x, src, left, top, width, height, rot = 0)
}
\arguments{
\item{x}{an pptx object}

\item{src}{image filename, the basename of the file must not contain any blank.}

\item{left, top}{location of the new shape on the slide}

\item{width, height}{image size in inches}

\item{rot}{rotation angle}
}
\description{
add an image as a new shape in the current slide.
This function is deprecated in favor of \code{\link{ph_with}}.
}
