% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_placeholder.R
\name{placeholder_add_paragraph}
\alias{placeholder_add_paragraph}
\title{append text}
\usage{
placeholder_add_paragraph(x, uid, level = 1)
}
\arguments{
\item{x}{a pptx device}

\item{uid}{unique id, use column \code{id} of the
result of \code{\link{summary.pptx}}}

\item{level}{paragraph level}
}
\description{
append text into a paragraph of a pptx object
}
\examples{
library(magrittr)

fileout <- tempfile(fileext = ".pptx")
default_text <- fp_text(font.size = 0, bold = TRUE, color = "red")

doc <- pptx() \%>\%
  add_slide(layout = "Title and Content", master = "Office Theme") \%>\%
  placeholder_set_text(id = "title", str = "A title") \%>\%
  placeholder_set_text(id = "body", str = "A text")

id_body <- summary(doc)$id[2]

placeholder_add_paragraph(x = doc, uid = id_body, level = 2) \%>\%
  placeholder_add_text(str = "and another, ", uid = id_body,
    style = default_text ) \%>\%
  placeholder_add_paragraph(uid = id_body, level = 3) \%>\%
  placeholder_add_text(str = "and another!", uid = id_body,
    style = update(default_text, color = "blue")) \%>\%
  print(target = fileout)
}

