% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_placeholder.R
\name{placeholder_add_xml}
\alias{placeholder_add_xml}
\title{add a string as xml}
\usage{
placeholder_add_xml(x, value, id = "body", index = 1)
}
\arguments{
\item{x}{a pptx device}

\item{value}{a character}

\item{id}{placeholder id}

\item{index}{placeholder index. This is to be used when a placeholder id
is not unique in the current slide, e.g. two placeholders with id 'body'.}
}
\description{
add a string (valid xml) into a pptx object
}

