% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_placeholder.R
\name{placeholder_remove}
\alias{placeholder_remove}
\title{remove shape}
\usage{
placeholder_remove(x, id = "title", index = 1)
}
\arguments{
\item{x}{a pptx device}

\item{id}{placeholder id}

\item{index}{placeholder index. This is to be used when a placeholder id
is not unique in the current slide, e.g. two placeholders with id 'body'.}
}
\description{
remove a shape in a slide
}
\examples{
fileout <- tempfile(fileext = ".pptx")
doc <- pptx()
doc <- add_slide(doc, layout = "Title and Content", master = "Office Theme")
doc <- placeholder_set_text(x = doc, id = "title", str = "Un titre")
doc <- placeholder_remove(x = doc, id = "title")
print(doc, target = fileout )
}

