% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_placeholder.R
\name{placeholder_set}
\alias{placeholder_set}
\title{add a new shape}
\usage{
placeholder_set(x, type = "title", index = 1)
}
\arguments{
\item{x}{a pptx device}

\item{type}{placeholder type}

\item{index}{placeholder index (integer). This is to be used when a placeholder type
is not unique in the current slide, e.g. two placeholders with type 'body'.}
}
\description{
add a new shape in a slide
}
\examples{
fileout <- tempfile(fileext = ".pptx")
doc <- read_pptx()
doc <- add_slide(doc, layout = "Title and Content", master = "Office Theme")
doc <- placeholder_set(x = doc, type = "title")

print(doc, target = fileout )
}

