% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_placeholder.R
\name{placeholder_set_img}
\alias{placeholder_set_img}
\title{add image into a new shape}
\usage{
placeholder_set_img(x, src, id, index = 1, width = NULL, height = NULL)
}
\arguments{
\item{x}{a pptx device}

\item{src}{image path}

\item{id}{placeholder id}

\item{index}{placeholder index. This is to be used when a placeholder id
is not unique in the current slide, e.g. two placeholders with id 'body'.}

\item{width, height}{image size in inches}
}
\description{
add an image into a new shape in a slide
}
\examples{
doc <- pptx()
doc <- add_slide(doc, layout = "Title and Content", master = "Office Theme")

img.file <- file.path( Sys.getenv("R_HOME"), "doc", "html", "logo.jpg" )
if( file.exists(img.file) ){
  doc <- placeholder_set_img(x = doc, id = "body", src = img.file, height = 1.06, width = 1.39 )
}
if( require("ionicons") ){
  calendar_src = as_png(name = "calendar", fill = "#FFE64D", width = 144, height = 144)
  doc <- placeholder_set_img(x = doc, id = "dt", src = calendar_src )
}

print(doc, target = "placeholder_set_img.pptx" )
}

