% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_placeholder.R
\name{placeholder_set_xml}
\alias{placeholder_set_xml}
\title{add a string as xml}
\usage{
placeholder_set_xml(x, value, type = "body", index = 1)
}
\arguments{
\item{x}{a pptx device}

\item{value}{a character}

\item{type}{placeholder type}

\item{index}{placeholder index (integer). This is to be used when a placeholder type
is not unique in the current slide, e.g. two placeholders with type 'body'.}
}
\description{
add a string (valid xml) into a pptx object
}

