% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_instr.R
\name{plot_instr}
\alias{plot_instr}
\title{Wrap plot instructions for png plotting in Powerpoint or Word}
\usage{
plot_instr(code)
}
\arguments{
\item{code}{plotting instructions}
}
\description{
A simple wrapper to capture
plot instructions that will be executed and copied in a document. It produces
an object of class 'plot_instr' with a corresponding method \code{\link[=ph_with]{ph_with()}}.

The function enable usage of any R plot with argument \code{code}. Wrap your code
between curly bracket if more than a single expression.
}
\examples{
anyplot <- plot_instr(code = barplot(1:5, col = 2:6))

library(officer)
doc <- read_pptx()
doc <- add_slide(doc, "Title and Content", "Office Theme")
doc <- ph_with(doc, anyplot, location = ph_location_fullsize())
print(doc, target = tempfile(fileext = ".pptx"))
}
\seealso{
\code{\link{ph_with}}
}
