% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx.R
\name{pptx}
\alias{pptx}
\alias{print.pptx}
\title{open a connexion to a 'PowerPoint' file}
\usage{
pptx(path = NULL)

\method{print}{pptx}(x, target = NULL, ...)
}
\arguments{
\item{path}{path to the pptx file to use a base document.}

\item{x}{a pptx object}

\item{target}{path to the pptx file to write}

\item{...}{unused}
}
\description{
read and import a pptx file as an R object
representing the document.
}
\examples{
# create a pptx object with default template ---
pptx()

}

