% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx_slide_manip.R
\name{pptx_fortify_slides}
\alias{pptx_fortify_slides}
\title{Correct pptx content references}
\usage{
pptx_fortify_slides(x)
}
\arguments{
\item{x}{an rpptx object}
}
\value{
an rpptx object
}
\description{
Content references are not managed directly
but computed after the content is added. This function
loop over each slide and fix references (links and images)
if necessary.
}
\seealso{
Other functions for officer extensions: 
\code{\link{docx_reference_img}()},
\code{\link{fortify_location}()},
\code{\link{get_reference_value}()},
\code{\link{opts_current_table}()},
\code{\link{shape_properties_tags}()},
\code{\link{to_html}()},
\code{\link{to_pml}()},
\code{\link{to_wml}()},
\code{\link{wml_link_images}()}
}
\concept{functions for officer extensions}
\keyword{internal}
