% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_border.R
\name{pr_border}
\alias{pr_border}
\alias{update.pr_border}
\title{border properties object}
\usage{
pr_border(color = "black", style = "solid", width = 1)

\method{update}{pr_border}(object, color, style, width, ...)
}
\arguments{
\item{color}{border color - single character value (e.g. "#000000" or "black")}

\item{style}{border style - single character value : "none" or "solid" or "dotted" or "dashed"}

\item{width}{border width - an integer value : 0>= value}

\item{object}{\code{pr_border} object to modify}

\item{...}{further arguments - not used}
}
\description{
create a border properties object.
}
\examples{
pr_border()
pr_border(color="orange", style="solid", width=1)
pr_border(color="gray", style="dotted", width=1)

# modify object ------
border <- pr_border()
update(border, style="dotted", width=3)
}

