% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_cell.R
\name{pr_cell}
\alias{format.pr_cell}
\alias{pr_cell}
\alias{print.pr_cell}
\alias{update.pr_cell}
\title{Cell formatting properties}
\usage{
pr_cell(border = pr_border(width = 0), border.bottom, border.left, border.top,
  border.right, vertical.align = "middle", margin = 0, margin.bottom,
  margin.top, margin.left, margin.right, background.color = "transparent",
  background.img.id = "rId1", background.img.src = NULL,
  text.direction = "lrtb", row_span = 1L, column_span = 1L)

\method{format}{pr_cell}(x, type = "wml", ...)

\method{print}{pr_cell}(x, ...)

\method{update}{pr_cell}(object, border, border.bottom, border.left, border.top,
  border.right, vertical.align, margin = 0, margin.bottom, margin.top,
  margin.left, margin.right, background.color, background.img.id,
  background.img.src, text.direction, row_span, column_span, ...)
}
\arguments{
\item{border}{shortcut for all borders.}

\item{border.bottom, border.left, border.top, border.right}{\code{\link{pr_border}} for borders.}

\item{vertical.align}{cell content vertical alignment - a single character value
, expected value is one of "center" or "top" or "bottom"}

\item{margin}{shortcut for all margins.}

\item{margin.bottom, margin.top, margin.left, margin.right}{cell margins - 0 or positive integer value.}

\item{background.color}{cell background color - a single character value specifying a
valid color (e.g. "#000000" or "black").}

\item{text.direction}{cell text rotation - a single character value, expected
value is one of "lrtb", "tbrl", "btlr".}

\item{row_span, column_span}{row and column span}

\item{x, object}{object \code{pr_cell}}

\item{type}{output type - one of 'wml', 'pml', 'html'.}

\item{...}{further arguments - not used}
}
\description{
Create a \code{pr_cell} object that describes cell formatting properties.
}
\examples{
obj <- pr_cell(margin = 1)
update( obj, margin.bottom = 5 )
}

