% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_par.R
\name{pr_par}
\alias{pr_par}
\alias{print.pr_par}
\alias{update.pr_par}
\title{Paragraph formatting properties}
\usage{
pr_par(text.align = "left", padding = 0, border = pr_border(width = 0),
  padding.bottom, padding.top, padding.left, padding.right, border.bottom,
  border.left, border.top, border.right, shading.color = "transparent")

\method{print}{pr_par}(x, ...)

\method{update}{pr_par}(object, text.align, padding, border, padding.bottom,
  padding.top, padding.left, padding.right, border.bottom, border.left,
  border.top, border.right, shading.color, ...)
}
\arguments{
\item{text.align}{text alignment - a single character value, expected value
is one of 'left', 'right', 'center', 'justify'.}

\item{padding}{paragraph paddings - 0 or positive integer value. Argument \code{padding} overwrites
arguments \code{padding.bottom}, \code{padding.top}, \code{padding.left}, \code{padding.right}.}

\item{border}{shortcut for all borders.}

\item{padding.bottom, padding.top, padding.left, padding.right}{paragraph paddings - 0 or positive integer value.}

\item{border.bottom, border.left, border.top, border.right}{\code{\link{pr_border}} for
borders. overwrite other border properties.}

\item{shading.color}{shading color - a single character value specifying
a valid color (e.g. "#000000" or "black").}

\item{x, object}{\code{pr_par} object}

\item{...}{further arguments - not used}
}
\value{
a \code{pr_par} object
}
\description{
Create a \code{pr_par} object that describes
paragraph formatting properties.
}
\details{
pr_par is used to control paragraph properties.
It is used when adding plots or when adding content in a FlexTable.
}
\examples{
pr_par(text.align = "center", padding = 5)
obj <- pr_par(text.align = "center", padding = 1)
update( obj, padding.bottom = 5 )
}

