% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_text.R
\name{pr_text}
\alias{format.pr_text}
\alias{pr_text}
\alias{print.pr_text}
\alias{update.pr_text}
\title{Text formatting properties}
\usage{
pr_text(color = "black", font.size = 10, bold = FALSE, italic = FALSE,
  underlined = FALSE, font.family = "Arial", vertical.align = "baseline",
  shading.color = "transparent")

\method{format}{pr_text}(x, type = "wml", ...)

\method{print}{pr_text}(x, ...)

\method{update}{pr_text}(object, color, font.size, bold = FALSE,
  italic = FALSE, underlined = FALSE, font.family, vertical.align,
  shading.color, ...)
}
\arguments{
\item{color}{font color - a single character value specifying
a valid color (e.g. "#000000" or "black").}

\item{font.size}{font size (in point) - 0 or positive integer value.}

\item{bold}{is bold}

\item{italic}{is italic}

\item{underlined}{is underlined}

\item{font.family}{single character value specifying font name.}

\item{vertical.align}{single character value specifying font vertical alignments.
Expected value is one of the following : default \code{'baseline'}
or \code{'subscript'} or \code{'superscript'}}

\item{shading.color}{shading color - a single character value specifying
a valid color (e.g. "#000000" or "black").}

\item{x}{\code{pr_text} object to print}

\item{type}{output type - one of 'wml', 'pml', 'html'.}

\item{...}{further arguments - not used}

\item{object}{\code{pr_text} object to modify}

\item{format}{format type, wml for MS word, pml for
MS PowerPoint and html.}
}
\value{
a \code{pr_text} object
}
\description{
Create a \code{pr_text} object that describes
text formatting properties.
}
\examples{
print( pr_text (color="red", font.size = 12) )
}

