% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf.R
\name{print.rtf}
\alias{print.rtf}
\title{Write an 'RTF' document to a file}
\usage{
\method{print}{rtf}(x, target = NULL, ...)
}
\arguments{
\item{x}{an 'rtf' object created with \code{\link[=rtf_doc]{rtf_doc()}}}

\item{target}{path to the RTF file to write}

\item{...}{unused}
}
\description{
Write the RTF object and its content to a file.
}
\examples{
# write a rdocx object in a rtf file ----
doc <- rtf_doc()
print(doc, target = tempfile(fileext = ".rtf"))
}
\seealso{
\code{\link[=rtf_doc]{rtf_doc()}}
}
