% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_block_objects.R
\name{prop_table}
\alias{prop_table}
\title{Table properties}
\usage{
prop_table(
  style = NA_character_,
  layout = table_layout(),
  width = table_width(),
  stylenames = table_stylenames(),
  colwidths = table_colwidths(),
  tcf = table_conditional_formatting(),
  align = "center",
  word_title = NULL,
  word_description = NULL
)
}
\arguments{
\item{style}{table style to be used to format table}

\item{layout}{layout defined by \code{\link[=table_layout]{table_layout()}},}

\item{width}{table width in the document defined by \code{\link[=table_width]{table_width()}}}

\item{stylenames}{columns styles defined by \code{\link[=table_stylenames]{table_stylenames()}}}

\item{colwidths}{column widths defined by \code{\link[=table_colwidths]{table_colwidths()}}}

\item{tcf}{conditional formatting settings defined by \code{\link[=table_conditional_formatting]{table_conditional_formatting()}}}

\item{align}{table alignment (one of left, center or right)}

\item{word_title}{alternative text for Word table (used as title of the table)}

\item{word_description}{alternative text for Word table (used as description of the table)}
}
\description{
Define table properties such as fixed or autofit layout,
table width in the document, eventually column widths.
}
\examples{
prop_table()
to_wml(prop_table())
}
\seealso{
Other functions for table definition:
\code{\link[=table_colwidths]{table_colwidths()}},
\code{\link[=table_conditional_formatting]{table_conditional_formatting()}},
\code{\link[=table_header]{table_header()}},
\code{\link[=table_layout]{table_layout()}},
\code{\link[=table_stylenames]{table_stylenames()}},
\code{\link[=table_width]{table_width()}}
}
\concept{functions for table definition}
