% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{read_clipboard}
\alias{from_clipboard}
\alias{read_clipboard}
\title{Read from Windows/OSX clipboard}
\usage{
read_clipboard(delim = "\\t", ...)

from_clipboard(delim = "\\t", ...)
}
\arguments{
\item{delim}{The delimiter for columns.}

\item{...}{Further arguments passed to \code{\link[readr]{read_delim}}.}
}
\description{
Thin wrapper for reading from windows/OSX clipboards with the most-used defaults.
The function first reads in the lines, checks if the delimiter is present in the lines
and then converts it to a data.frame.
}
\note{
This function only works on Windows or OSX, and the data-size cannot
exceed 128kb in Windows.
}
\examples{
\dontrun{
# Only works on Windows and OSX
df <- data.frame("String" = c("A", "B"), "Int" = c(1:2L), "Percent" = c(0.5, 0.75))
to_clipboard(df)
x <- from_clipboard()

# All equal - except attributes
# (readr attaches attr "problems")
all.equal(x, df, check.attributes = FALSE)
}
}
\author{
Kristian D. Olsen
}

