% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_read_styles.R
\name{read_core}
\alias{read_core}
\title{core properties}
\usage{
read_core(path)
}
\arguments{
\item{path}{path to Word document}
}
\description{
read core properties of a Word file and returns a tabular description
}
\examples{
template_dir <- tempfile()
docx_file <- system.file(package = "officer", "template/template.docx")
unzip( zipfile = docx_file, exdir = template_dir )
read_core( path = template_dir )
}
