% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_relationship.R
\name{read_relationship}
\alias{read_relationship}
\title{read openxml relationship content}
\usage{
read_relationship(x)
}
\arguments{
\item{x}{an xml string or a file name}
}
\description{
read a string or a file from a relationship file and return a detailed
data.frame containing imported attributes of the relationship file. These relationship
files are located in openxml packages.
}

