% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_read_styles.R
\name{read_styles}
\alias{read_styles}
\title{read word styles}
\usage{
read_styles(path)
}
\arguments{
\item{path}{path to Word document}
}
\description{
read word styles and returns a tabular description
}
\examples{
template_dir <- tempfile()
docx_file <- system.file(package = "officer", "template/template.docx")
unzip( zipfile = docx_file, exdir = template_dir )
read_styles( path = template_dir )
}

