% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx_slide_manip.R
\name{remove_slides_bulk}
\alias{remove_slides_bulk}
\title{Remove multiple slides at once}
\usage{
remove_slides_bulk(x, indices, rm_images = TRUE)
}
\arguments{
\item{x}{an rpptx object}

\item{indices}{a numeric vector of slide indices to remove}

\item{rm_images}{if TRUE (defaults to TRUE), images presented in
the slides to remove are also removed from the file.}
}
\description{
Remove multiple slides from a PowerPoint presentation in a single operation.
This is more efficient than calling \code{\link{remove_slide}} multiple times.
}
\note{
cursor is set on the last slide after removal.
}
\examples{
# Create a presentation with multiple slides
my_pres <- read_pptx()
my_pres <- add_slide(my_pres, "Title and Content")
my_pres <- add_slide(my_pres, "Title and Content") 
my_pres <- add_slide(my_pres, "Title and Content")
my_pres <- add_slide(my_pres, "Title and Content")

# Remove slides 2 and 4
my_pres <- remove_slides_bulk(my_pres, c(2, 4))
}
\seealso{
\code{\link[=read_pptx]{read_pptx()}}, \code{\link[=remove_slide]{remove_slide()}}, \code{\link[=ph_with]{ph_with()}}

Other functions to manipulate slides:
\code{\link{add_slide}()},
\code{\link{clean_unused_media}()},
\code{\link{move_slide}()},
\code{\link{on_slide}()},
\code{\link{remove_slide}()},
\code{\link{set_notes}()}
}
\concept{slide_manipulation}
