% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_run_objects.R
\name{run_bookmark}
\alias{run_bookmark}
\title{bookmark for Word}
\usage{
run_bookmark(bkm, run)
}
\arguments{
\item{bkm}{bookmark id to associate with run. Value can only be
made of alpha numeric characters, '-' and '_'.}

\item{run}{a run object, made with a call to one of
the "run functions for reporting".}
}
\description{
Add a bookmark on a run object.
}
\section{usage}{

You can use this function in conjunction with \link{fpar} to create paragraphs
consisting of differently formatted text parts. You can also use this
function as an \emph{r chunk} in an R Markdown document made with package
officedown.
}

\examples{
ft <- fp_text(font.size = 12, bold = TRUE)
run_bookmark("par1", ftext("some text", ft))
}
\seealso{
Other run functions for reporting: 
\code{\link{external_img}()},
\code{\link{ftext}()},
\code{\link{hyperlink_ftext}()},
\code{\link{run_autonum}()},
\code{\link{run_columnbreak}()},
\code{\link{run_linebreak}()},
\code{\link{run_pagebreak}()},
\code{\link{run_reference}()},
\code{\link{run_word_field}()}
}
\concept{run functions for reporting}
