% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_run_objects.R
\name{run_columnbreak}
\alias{run_columnbreak}
\title{column break}
\usage{
run_columnbreak()
}
\description{
Create a representation of a column break
}
\examples{
run_columnbreak()
}
\seealso{
Other run functions for reporting: 
\code{\link{external_img}()},
\code{\link{ftext}()},
\code{\link{run_autonum}()},
\code{\link{run_linebreak}()},
\code{\link{run_pagebreak}()},
\code{\link{run_reference}()},
\code{\link{run_seqfield}()}
}
\concept{run functions for reporting}
