% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_run_objects.R
\name{run_linebreak}
\alias{run_linebreak}
\title{page break for Word}
\usage{
run_linebreak()
}
\description{
Object representing a line break for a Word document. The
result must be used within a call to \link{fpar}.
}
\examples{

fp_t <- fp_text(font.size = 12, bold = TRUE)
an_fpar <- fpar("let's add a line break", run_linebreak(), ftext("and blah blah!", fp_t))

x <- read_docx()
x <- body_add(x, an_fpar)
print(x, target = tempfile(fileext = ".docx"))
}
\seealso{
Other run functions for reporting: 
\code{\link{external_img}()},
\code{\link{ftext}()},
\code{\link{run_pagebreak}()}
}
\concept{run functions for reporting}
