% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_run_objects.R
\name{run_pagebreak}
\alias{run_pagebreak}
\title{page break for Word}
\usage{
run_pagebreak()
}
\description{
Object representing a page break for a Word document. The
result must be used within a call to \link{fpar}.
}
\examples{

fp_t <- fp_text(font.size = 12, bold = TRUE)
an_fpar <- fpar("let's add a break page", run_pagebreak(), ftext("and blah blah!", fp_t))

x <- read_docx()
x <- body_add(x, an_fpar)
print(x, target = tempfile(fileext = ".docx"))
}
\seealso{
Other run functions for reporting: 
\code{\link{external_img}()},
\code{\link{ftext}()},
\code{\link{run_autonum}()},
\code{\link{run_columnbreak}()},
\code{\link{run_linebreak}()},
\code{\link{run_reference}()},
\code{\link{run_seqfield}()}
}
\concept{run functions for reporting}
