% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_run_objects.R
\name{run_reference}
\alias{run_reference}
\title{reference}
\usage{
run_reference(id)
}
\arguments{
\item{id}{reference id, a string}
}
\description{
Create a representation of a reference
}
\examples{
run_reference('a_ref')
}
