% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_run_objects.R
\name{run_word_field}
\alias{run_word_field}
\title{'Word' computed field}
\usage{
run_word_field(field, prop = NULL)
}
\arguments{
\item{field}{Value for a "Word Computed Field" as a string.}

\item{prop}{formatting text properties returned by \link{fp_text}.}
}
\description{
Create a 'Word' computed field.
}
\section{usage}{

You can use this function in conjunction with \link{fpar} to create paragraphs
consisting of differently formatted text parts. You can also use this
function as an \emph{r chunk} in an R Markdown document made with package
officedown.
}

\examples{
run_word_field(field = "PAGE  \\\\* MERGEFORMAT")
run_word_field(field = "Date \\\\@ \"MMMM d yyyy\"")
}
\seealso{
Other run functions for reporting:
\code{\link[=external_img]{external_img()}},
\code{\link[=floating_external_img]{floating_external_img()}},
\code{\link[=ftext]{ftext()}},
\code{\link[=hyperlink_ftext]{hyperlink_ftext()}},
\code{\link[=run_autonum]{run_autonum()}},
\code{\link[=run_bookmark]{run_bookmark()}},
\code{\link[=run_columnbreak]{run_columnbreak()}},
\code{\link[=run_comment]{run_comment()}},
\code{\link[=run_footnote]{run_footnote()}},
\code{\link[=run_footnoteref]{run_footnoteref()}},
\code{\link[=run_linebreak]{run_linebreak()}},
\code{\link[=run_pagebreak]{run_pagebreak()}},
\code{\link[=run_reference]{run_reference()}},
\code{\link[=run_tab]{run_tab()}},
\code{\link[=run_wordtext]{run_wordtext()}}

Other Word computed fields:
\code{\link[=run_autonum]{run_autonum()}},
\code{\link[=run_reference]{run_reference()}}
}
\concept{Word computed fields}
\concept{run functions for reporting}
