% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-proc.R
\name{sanitize_images}
\alias{sanitize_images}
\title{Remove unused media from a document}
\usage{
sanitize_images(x, warn_user = TRUE)
}
\arguments{
\item{x}{\code{rdocx} or \code{rpptx} object}

\item{warn_user}{TRUE to make sure users are warned when
using this function that will be un-exported in a next
version}
}
\description{
The function will scan the media
directory and delete images that are not used
anymore. This function is to be used when images
have been replaced many times.
}
\keyword{internal}
