% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seamless.R
\docType{package}
\name{seamless}
\alias{seamless}
\alias{seamless-package}
\title{seamless: A toolbox for seamlessly integrating R}
\description{
This package is meant to provide a more "seamless" experience when using
R in a workflow with MS Office and other programs in general. See links in
details for more information.
}
\details{
\itemize{
\item Get data in and out of R with \code{\link{read_data}} and \code{\link{write_data}}.
\item Access sharepoint using \code{\link{sharepoint_link}} and/or \code{\link{sharepoint_mount}}.
\item Read/write to the clipboard on Windows and Os X using \code{\link{from_clipboard}} and \code{\link{to_clipboard}}.
\item Send data or plots to Excel and Powerpoint with \code{\link{to_excel}} and \code{\link{to_ppt}}.
}
}
\author{
Kristian D. Olsen
}

