% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_run_objects.R
\name{section_columns}
\alias{section_columns}
\title{Section columns}
\usage{
section_columns(widths = c(2.5, 2.5), space = 0.25, sep = FALSE)
}
\arguments{
\item{widths}{columns widths in inches. If 3 values, 3 columns
will be produced.}

\item{space}{space in inches between columns.}

\item{sep}{if TRUE a line is separating columns.}
}
\description{
The function creates a representation of the columns of a section.
}
\examples{
section_columns()
}
\seealso{
Other functions for section definition: 
\code{\link{page_mar}()},
\code{\link{page_size}()},
\code{\link{prop_section}()}
}
\concept{functions for section definition}
