% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_section.R
\name{sections}
\alias{sections}
\alias{body_end_section_continuous}
\alias{body_end_section_landscape}
\alias{body_end_section_portrait}
\alias{body_end_section_columns}
\alias{body_end_section_columns_landscape}
\title{section}
\usage{
body_end_section_continuous(x)

body_end_section_landscape(x, w = 21/2.54, h = 29.7/2.54)

body_end_section_portrait(x, w = 21/2.54, h = 29.7/2.54)

body_end_section_columns(x, widths = c(2.5, 2.5), space = 0.25,
  sep = FALSE)

body_end_section_columns_landscape(x, widths = c(2.5, 2.5), space = 0.25,
  sep = FALSE, w = 21/2.54, h = 29.7/2.54)
}
\arguments{
\item{x}{an rdocx object}

\item{w, h}{width and height in inches of the section page. This will
be ignored if the default section (of the \code{reference_docx} file)
already has width and height.}

\item{widths}{columns widths in inches. If 3 values, 3 columns
will be produced.}

\item{space}{space in inches between columns.}

\item{sep}{if TRUE a line is separating columns.}
}
\description{
Add a section break.
}
\details{
A section starts at the end of the previous section (or the beginning of
the document if no preceding section exists), and stops where the section is declared.
}
