% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_add.R
\name{seqfield}
\alias{seqfield}
\title{docx seqfield xml}
\usage{
seqfield(sf, as_scalar = TRUE)
}
\arguments{
\item{sf}{seqfield value}
}
\description{
Create a string representation of a seqfield.
}
\examples{
seqfield(" REF label \\\\h ")
}
