% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{set_title}
\alias{set_title}
\title{add a title}
\usage{
set_title(x, value)
}
\arguments{
\item{x}{a pptx device}

\item{value}{a character}

\item{pos}{where to add the new element relative to the cursor,
one of "after", "before", "on".}
}
\description{
add a title into a pptx object
}

