% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint.R
\name{sharepoint_link}
\alias{read_data.mounted_sharepoint}
\alias{read_data.sharepoint}
\alias{sharepoint_link}
\title{Prepare a sharepoint link for reading}
\usage{
sharepoint_link(link, mounted = FALSE)

\method{read_data}{sharepoint}(x, destination = NULL, ...)

\method{read_data}{mounted_sharepoint}(x, ...)
}
\arguments{
\item{link}{A string starting with \code{http}.}

\item{mounted}{Set this to \code{TRUE} if you are on windows and have mounted
the sharepoint server as a network drive.}

\item{x}{An object returned from \code{sharepoint_link}.}

\item{destination}{Optional: When sharepoint is not mounted, files are always
downloaded before being read into R. The file is temporary if you do not set
the destination.}

\item{...}{Further arguments passed to \code{read_data}.}
}
\description{
This function is minimalistic wrapper for \code{httr}, with the intention
of making it easy to read files from sharepoint using \code{read_data}.
}
\note{
This function requires \code{httr} if the drive is not mounted. User/password
will be requested and stored for the current R session. Imporant that you don't
share your .Rhistory after this!
}
\examples{
\dontrun{
# The gist of it:
read_data(sharepoint_link(your_server))
}
}
\author{
Kristian D. Olsen
}

