% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint.R
\name{sharepoint_mount}
\alias{read_data.sharepoint_link}
\alias{sharepoint_link}
\alias{sharepoint_mount}
\title{Prepare a sharepoint link for reading}
\usage{
sharepoint_mount(link)

sharepoint_link(link)

\method{read_data}{sharepoint_link}(file, destination = NULL, ...)
}
\arguments{
\item{link}{A string starting with \code{http}.}

\item{file}{An object returned from \code{sharepoint_link}.}

\item{destination}{Optional: When sharepoint is not mounted, files are always
downloaded before being read into R. The file is temporary if you do not set
the destination.}

\item{...}{Further arguments passed to \code{\link{read_data}}.}
}
\description{
This function is minimalistic wrapper for \pkg{httr}, with the intention
of making it easy to read files from sharepoint using \code{\link{read_data}}.
}
\note{
This function requires \pkg{httr} if the drive is not mounted. User/password
will be requested and stored for the current R session. Imporant that you don't
share your .Rhistory after this!
}
\examples{
\dontrun{
# The gist of it:
read_data(sharepoint_link(your_server))
read_data(sharepoint_mount(your_server))
}
}
\author{
Kristian D. Olsen
}

