% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xlsx.R
\name{sheet_add_table}
\alias{sheet_add_table}
\title{add a data.frame}
\usage{
sheet_add_table(x, sheet, value, at_row = 1, at_col = 1)
}
\arguments{
\item{x}{rxlsx object}

\item{sheet}{sheet label/name}

\item{value}{data.frame}

\item{at_row, at_col}{where to start writing}
}
\description{
add a data.frame into an xlsx worksheet
}
\examples{
my_ws <- read_xlsx()
my_pres <- add_sheet(my_ws, label = "new sheet")
my_pres <- sheet_add_table(my_ws, sheet = "new sheet", iris)
}
