% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xlsx.R
\name{sheet_names}
\alias{sheet_names}
\title{Sheet names of an xlsx workbook}
\usage{
sheet_names(x)
}
\arguments{
\item{x}{an rxlsx object (created by \code{\link[=read_xlsx]{read_xlsx()}}).}
}
\value{
A character vector of sheet names.
}
\description{
Return the sheet names of an \code{rxlsx} object, in the
order they appear in the workbook.
}
\examples{
wb <- read_xlsx()
sheet_names(wb)

wb <- add_sheet(wb, label = "new sheet")
sheet_names(wb)
}
