% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xlsx.R
\name{sheet_view}
\alias{sheet_view}
\title{select sheet selected}
\usage{
sheet_view(x, sheet)
}
\arguments{
\item{x}{rxlsx object}

\item{sheet}{sheet name}
}
\description{
set a particular sheet selected when workbook will be
edited.
}
\examples{
my_ws <- read_xlsx()
my_pres <- add_sheet(my_ws, label = "new sheet")
my_pres <- sheet_view(my_ws, sheet = "new sheet")
print(my_ws, target = tempfile(fileext = ".xlsx") )
}
