% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xlsx.R
\name{sheet_write_data}
\alias{sheet_write_data}
\title{Write data to a sheet}
\usage{
sheet_write_data(x, data, sheet, start_row = 1L, start_col = 1L)
}
\arguments{
\item{x}{rxlsx object}

\item{data}{a data.frame}

\item{sheet}{sheet name (must already exist)}

\item{start_row}{row index where the header will be written (default 1)}

\item{start_col}{column index where the first column of data will be
written (default 1)}
}
\description{
Write a data.frame into a sheet of an xlsx workbook.
Multiple calls can write to different positions on the same sheet.
}
\examples{
x <- read_xlsx()
x <- add_sheet(x, label = "mysheet")
x <- sheet_write_data(x, data = head(iris, 5), sheet = "mysheet")
x <- sheet_write_data(x, data = head(mtcars, 3), sheet = "mysheet",
  start_col = 7)
print(x, target = tempfile(fileext = ".xlsx"))
}
