% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorcuts.R
\name{shortcuts}
\alias{shortcuts}
\title{shortcuts for formatting properties}
\usage{
shortcuts
}
\description{
Shortcuts for \code{\link[=fp_text]{fp_text()}}, \code{\link[=fp_par]{fp_par()}}, \code{\link[=fp_cell]{fp_cell()}} and \code{\link[=fp_border]{fp_border()}}.
}
\examples{
shortcuts$fp_bold()
shortcuts$fp_italic()
shortcuts$b_null()
}
