% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_section.R
\name{break_column_before}
\alias{break_column_before}
\alias{slip_in_column_break}
\title{add a column break}
\usage{
break_column_before(x)

slip_in_column_break(x, pos = "before")
}
\arguments{
\item{x}{an rdocx object}

\item{pos}{where to add the new element relative to the cursor,
"after" or "before".}
}
\description{
add a column break into a Word document. A column break
is used to add a break in a multi columns section in a Word
Document.
}
