% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_footnotes.R
\name{slip_in_footnote}
\alias{slip_in_footnote}
\title{append a footnote}
\usage{
slip_in_footnote(x, style = NULL, blocks, pos = "after")
}
\arguments{
\item{x}{an rdocx object}

\item{style}{text style to be used for the reference note}

\item{blocks}{set of blocks to be used as footnote content returned by
function \code{\link{block_list}}.}

\item{pos}{where to add the new element relative to the cursor, "after" or
"before".}
}
\description{
append a new footnote into a paragraph of an rdocx object
}
\note{
This function will be deprecated in the next release because it is not
efficient and make users write complex code. Use instead \code{\link[=fpar]{fpar()}} to build
formatted paragraphs.
}
