% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{slip_in_img}
\alias{slip_in_img}
\title{append an image}
\usage{
slip_in_img(x, src, style = NULL, width, height, pos = "after")
}
\arguments{
\item{x}{an rdocx object}

\item{src}{image filename, the basename of the file must not contain any blank.}

\item{style}{text style}

\item{width}{height in inches}

\item{height}{height in inches}

\item{pos}{where to add the new element relative to the cursor,
"after" or "before".}
}
\description{
append an image into a paragraph of an rdocx object.

This function will be deprecated in the next release because it is not
efficient and make users write complex code. Use instead \code{\link[=fpar]{fpar()}} to build
formatted paragraphs.
}
