% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_add_chunk.R
\name{slip_in_xml}
\alias{slip_in_xml}
\title{add a wml string into a Word document}
\usage{
slip_in_xml(x, str, pos)
}
\arguments{
\item{x}{a docx object}

\item{str}{a wml string}

\item{pos}{where to add the new element relative to the cursor,
"after" or "before".}
}
\description{
The function add a wml string into
the document after, before or on a cursor location.
}
