% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cot_objects.R
\name{stext}
\alias{stext}
\title{styled text}
\usage{
stext(text, id)
}
\arguments{
\item{text}{text value}

\item{id}{style name}
}
\description{
Format a chunk of text with a referenced style.
This will produce a simple text when used in a pptx presentation
as the style concept can not be apply for this format.
}
\examples{
stext("hello", "strong")
}
