% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop_if_not_in_slide_range}
\alias{stop_if_not_in_slide_range}
\title{Ensure valid slide indexes}
\usage{
stop_if_not_in_slide_range(x, idx, arg = NULL, call = parent.frame())
}
\arguments{
\item{x}{An \code{rpptx} object.}

\item{idx}{Slide indexes.}

\item{arg}{Name of argument to use in error message (optional).}

\item{call}{Environment to display in error message. Defaults to caller env.
Set \code{NULL} to suppress (see \link[cli:cli_abort]{cli::cli_abort}).}
}
\description{
Ensure valid slide indexes
}
\keyword{internal}
