% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf.R
\name{str_encode_to_rtf}
\alias{str_encode_to_rtf}
\title{Encode UTF8 string to RTF}
\usage{
str_encode_to_rtf(z)
}
\arguments{
\item{z}{character vector to be converted}
}
\value{
character vector of results encoded to RTF
}
\description{
Convert strings to RTF valid codes.
}
\examples{
str_encode_to_rtf("Hello World")
}
\seealso{
Other functions for officer extensions: 
\code{\link{docx_reference_img}()},
\code{\link{fortify_location}()},
\code{\link{get_reference_value}()},
\code{\link{opts_current_table}()},
\code{\link{shape_properties_tags}()},
\code{\link{to_html}()},
\code{\link{to_pml}()},
\code{\link{to_rtf}()},
\code{\link{to_wml}()},
\code{\link{wml_link_images}()}
}
\concept{functions for officer extensions}
\keyword{internal}
