% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_docx.R
\name{styles_info}
\alias{styles_info}
\title{read Word styles}
\usage{
styles_info(
  x,
  type = c("paragraph", "character", "table", "numbering"),
  is_default = c(TRUE, FALSE)
)
}
\arguments{
\item{x}{an rdocx object}

\item{type, is_default}{subsets for types (i.e. paragraph) and
default style (when \code{is_default} is TRUE or FALSE)}
}
\description{
read Word styles and get results in
a data.frame.
}
\examples{
x <- read_docx()
styles_info(x)
styles_info(x, type = "paragraph", is_default = TRUE)
}
\seealso{
Other functions for Word document informations: 
\code{\link{doc_properties}()},
\code{\link{docx_bookmarks}()},
\code{\link{docx_dim}()},
\code{\link{length.rdocx}()},
\code{\link{set_doc_properties}()}
}
\concept{functions for Word document informations}
