% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_block_objects.R
\name{table_conditional_formatting}
\alias{table_conditional_formatting}
\title{Table conditional formatting}
\usage{
table_conditional_formatting(
  first_row = TRUE,
  first_column = FALSE,
  last_row = FALSE,
  last_column = FALSE,
  no_hband = FALSE,
  no_vband = TRUE
)
}
\arguments{
\item{first_row, last_row}{apply or remove formatting from the first or last row in the table.}

\item{first_column, last_column}{apply or remove formatting from the first or last column in the table.}

\item{no_hband, no_vband}{don't display odd and even rows or columns with
alternating shading for ease of reading.}
}
\description{
Tables can be conditionally formatted based on few properties as
whether the content is in the first row, last row, first column, or last
column, or whether the rows or columns are to be banded.
}
\note{
You must define a format for first_row, first_column and other properties
if you need to use them. The format is defined in a docx template.
}
\examples{
table_conditional_formatting(first_row = TRUE, first_column = TRUE)
}
\seealso{
Other functions for table definition:
\code{\link[=prop_table]{prop_table()}},
\code{\link[=table_colwidths]{table_colwidths()}},
\code{\link[=table_header]{table_header()}},
\code{\link[=table_layout]{table_layout()}},
\code{\link[=table_stylenames]{table_stylenames()}},
\code{\link[=table_width]{table_width()}}
}
\concept{functions for table definition}
