% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_block_objects.R
\name{table_layout}
\alias{table_layout}
\title{Algorithm for table layout}
\usage{
table_layout(type = "autofit")
}
\arguments{
\item{type}{'autofit' or 'fixed' algorithm. Default to 'autofit'.}
}
\description{
When a table is displayed in a document, it can
either be displayed using a fixed width or autofit layout algorithm:
\itemize{
\item fixed: uses fixed widths for columns. The width of the table is not
changed regardless of the contents of the cells.
\item autofit: uses the contents of each cell and the table width to
determine the final column widths.
}
}
\seealso{
Other functions for table definition:
\code{\link[=prop_table]{prop_table()}},
\code{\link[=table_colwidths]{table_colwidths()}},
\code{\link[=table_conditional_formatting]{table_conditional_formatting()}},
\code{\link[=table_header]{table_header()}},
\code{\link[=table_stylenames]{table_stylenames()}},
\code{\link[=table_width]{table_width()}}
}
\concept{functions for table definition}
