% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_run_objects.R
\name{to_html}
\alias{to_html}
\title{Convert officer objects to HTML}
\usage{
to_html(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{Arguments to be passed to methods}
}
\description{
Convert an object made with package officer
to HTML. The result is a string.
}
\seealso{
Other functions for officer extensions: 
\code{\link{docx_reference_img}()},
\code{\link{fortify_location}()},
\code{\link{get_reference_value}()},
\code{\link{opts_current_table}()},
\code{\link{shape_properties_tags}()},
\code{\link{to_pml}()},
\code{\link{to_wml}()},
\code{\link{wml_link_images}()}
}
\concept{functions for officer extensions}
\keyword{internal}
