% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_run_objects.R
\name{to_pml}
\alias{to_pml}
\title{Convert officer objects to PresentationML}
\usage{
to_pml(x, add_ns = FALSE, ...)
}
\arguments{
\item{x}{object}

\item{add_ns}{should namespace be added to the top tag}

\item{...}{Arguments to be passed to methods}
}
\description{
Convert an object made with package officer
to PresentationML. The result is a string.
}
