% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ppt.R
\name{to_ppt}
\alias{ppt_workbook}
\alias{to_ppt}
\alias{to_ppt.FlexTable}
\alias{to_ppt.character}
\alias{to_ppt.data.frame}
\alias{to_ppt.ggplot}
\title{Pass data to Powerpoint workbooks}
\usage{
to_ppt(x, wb, title = NULL, subtitle = NULL)

ppt_workbook(template = NULL, font = "Calibri", fontsize = 10L)

\method{to_ppt}{data.frame}(x, wb, title = NULL, subtitle = NULL)

\method{to_ppt}{FlexTable}(x, wb, title = NULL, subtitle = NULL)

\method{to_ppt}{ggplot}(x, wb, title = NULL, subtitle = NULL)

\method{to_ppt}{character}(x, wb, title = NULL, subtitle = NULL)
}
\arguments{
\item{x}{A \code{data.frame}, \code{string} or \code{plot}}

\item{wb}{A \code{pptWorkbook}.}

\item{title}{Title to use for the new slide.}

\item{subtitle}{Subtitle.}

\item{template}{Optional: Specify a path if you would like to append slides to
an existing powerpoint file (use it as a template). Default uses the template
included in this package.}

\item{font}{Default font (Replaces the \code{ReporteRs-default-font} option).}

\item{fontsize}{Default fontsize (Replaces the \code{ReporteRs-fontsize} option).}
}
\description{
\code{to_ppt} allows you to pass R objects to an open \code{pptWorkbook},
and write it later with \code{write_data}. The \code{pptWorkbook} can be created
by calling \code{ppt_workbook}. \code{to_ppt} will always format
the output and add it to a new slide.
}
\details{
You can use this function to pass \code{plot}, \code{data.frame} (\code{table} and
\code{matrix} will be coerced), and \code{character} objects to the workbook.
Strings are assumed to be markdown.
}
\note{
This function requires \pkg{ReporteRs}. The \code{pptWorkbook} object
is a thin R6 wrapper around ReporteR's \code{pptx}, and allow us to use
\code{to_ppt} in chained expressions, since the workbook is mutable.
}
\examples{
if (require(ReporteRs)) {
 wb <- ppt_workbook()
 df <- data.frame("String" = c("A", "B"), "Int" = c(1:2L), "Percent" = c(0.5, 0.75))

 # The workbook is mutable, so we don't have to assign result.
 to_ppt(df, wb, title = "Example data", subtitle = "")

 # Data is first argument, so we can use it with dplyr.
 # df \%>\% to_ppt(wb, title = "Example data", subtitle = "")

 # Save the data
 write_data(wb, "Example table.pptx")
}
}
\author{
Kristian D. Olsen
}

