% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_folder.R
\name{unpack_folder}
\alias{unpack_folder}
\title{Extract files from a zip file}
\usage{
unpack_folder(file, folder)
}
\arguments{
\item{file}{path of the archive to unzip}

\item{folder}{folder to create}
}
\description{
Extract files from a zip file to a folder. The
function returns the complete path to destination folder.
}
\keyword{internal}
