% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev-utils.R
\name{uuid_generate}
\alias{uuid_generate}
\title{generates unique identifiers}
\usage{
uuid_generate(n = 1, ...)
}
\arguments{
\item{n}{integer, number of unique identifiers to generate.}

\item{...}{arguments sent to \code{\link[=UUIDgenerate]{UUIDgenerate()}}}
}
\description{
generates unique identifiers based
on \code{\link[=UUIDgenerate]{UUIDgenerate()}}.
}
\examples{
uuid_generate(n = 5)
}
\keyword{internal}
