% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev-utils.R
\name{UUIDgenerate}
\alias{UUIDgenerate}
\title{generates pseudo unique identifiers}
\usage{
UUIDgenerate(n = 1, prefix = "", ...)
}
\arguments{
\item{n}{integer, number of unique identifiers to generate.}

\item{prefix}{a prefix to add}

\item{...}{unused arguments}
}
\description{
generates unique identifiers based
on \code{\link[=rand_bytes]{rand_bytes()}}.
}
\examples{
UUIDgenerate(n = 5)
UUIDgenerate(n = 5, prefix = "cl-")
}
\keyword{internal}
