% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_output.R
\name{write_clipboard}
\alias{to_clipboard}
\alias{write_clipboard}
\title{Write to Windows/OSX clipboard}
\usage{
write_clipboard(x, ...)

to_clipboard(x, ...)
}
\arguments{
\item{x}{The data or text to write.}

\item{...}{Further arguments passed to \code{\link{write.table}}.}
}
\description{
Wrapper for writing to windows/OSX clipboards with the most-used defaults for a
scandinavian locale.
}
\note{
This function only works on Windows or OSX, and the data-size cannot
exceed 128kb in Windows.
}
\examples{
\dontrun{
# Only works on Windows and OSX
df <- data.frame("String" = c("A", "B"), "Int" = c(1:2L), "Percent" = c(0.5, 0.75))
to_clipboard(df)
}
}
\author{
Kristian D. Olsen
}

