% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xlsx.R
\name{add_sheet}
\alias{add_sheet}
\title{Add a sheet}
\usage{
add_sheet(x, label)
}
\arguments{
\item{x}{rxlsx object}

\item{label}{sheet label}
}
\description{
Add a sheet into an xlsx worksheet.
}
\examples{
my_ws <- read_xlsx()
my_pres <- add_sheet(my_ws, label = "new sheet")
}
