% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx.R
\name{add_slide}
\alias{add_slide}
\title{add a slide}
\usage{
add_slide(x, layout = "Titre et contenu", master = "masque1")
}
\arguments{
\item{x}{pptx object}

\item{layout}{slide layout name to use}

\item{master}{master layout name where \code{layout} is located}
}
\description{
add a slide to a pptx presentation
}

