% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx_slide_manip.R
\name{add_slide}
\alias{add_slide}
\title{Add a slide}
\usage{
add_slide(x, layout = "Title and Content", master = "Office Theme")
}
\arguments{
\item{x}{an rpptx object}

\item{layout}{slide layout name to use}

\item{master}{master layout name where \code{layout} is located}
}
\description{
Add a slide into a pptx presentation.
}
\examples{
my_pres <- read_pptx()
layout_summary(my_pres)
my_pres <- add_slide(my_pres,
  layout = "Two Content", master = "Office Theme"
)
}
\seealso{
\code{\link[=print.rpptx]{print.rpptx()}}, \code{\link[=read_pptx]{read_pptx()}}, \code{\link[=plot_layout_properties]{plot_layout_properties()}}, \code{\link[=ph_with]{ph_with()}}, \code{\link[=layout_summary]{layout_summary()}}

Other functions slide manipulation: 
\code{\link{move_slide}()},
\code{\link{on_slide}()},
\code{\link{remove_slide}()},
\code{\link{set_notes}()}
}
\concept{functions slide manipulation}
