% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pptx.R
\name{add_slide}
\alias{add_slide}
\title{add a slide}
\usage{
add_slide(x, layout, master)
}
\arguments{
\item{x}{pptx object}

\item{layout}{slide layout name to use}

\item{master}{master layout name where \code{layout} is located}
}
\description{
add a slide into a pptx presentation
}
\examples{
my_pres <- read_pptx()
my_pres <- add_slide(my_pres,
  layout = "Two Content", master = "Office Theme")
}
