% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pptx.R
\name{annotate_base}
\alias{annotate_base}
\title{annotate PowerPoint base document}
\usage{
annotate_base(path = NULL, output_file = "annotated_layout.pptx")
}
\arguments{
\item{path}{path to the pptx file to use as base document or NULL to use the officer default}

\item{output_file}{filename to store the annotated powerpoint file or NULL to suppress generation}
}
\value{
x rpptx object of the annotated PowerPoint file
}
\description{
generates a slide from each layout in the base document to
identify the placeholder indexes, master names and indexes.
}
\examples{
# To generate an anotation of the default base document with officer:
annotate_base(output_file = tempfile(fileext = ".pptx"))

# To generate an annotation of the base document 'mydoc.pptx' and place the
# annotated output in 'mydoc_annotate.pptx'
# annotate_base(path = 'mydoc.pptx', output_file='mydoc_annotate.pptx')

}
