% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_block_objects.R
\name{block_caption}
\alias{block_caption}
\title{caption block}
\usage{
block_caption(label, style, id, autonum = NULL)
}
\arguments{
\item{label}{a scalar character representing label to display}

\item{style}{paragraph style name}

\item{id}{cross reference identifier}

\item{autonum}{an object generated with function \link{run_autonum}}
}
\description{
Create a representation of a
caption that can be used for cross reference. The caption
can also be an auto numbered paragraph.
}
\examples{
block_caption("a caption", style = "Normal", id = "caption_id")
block_caption("a caption",
  style = "Normal", id = "caption_id",
  autonum = run_autonum()
)
}
\seealso{
Other block functions for reporting: 
\code{\link{block_section}()},
\code{\link{block_table}()},
\code{\link{block_toc}()}
}
\concept{block functions for reporting}
