% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_block_objects.R
\name{block_table}
\alias{block_table}
\title{table}
\usage{
block_table(
  x,
  style = NULL,
  header = TRUE,
  first_row = TRUE,
  first_column = FALSE,
  last_row = FALSE,
  last_column = FALSE,
  no_hband = FALSE,
  no_vband = TRUE
)
}
\arguments{
\item{x}{a data.frame to add as a table}

\item{style}{table style}

\item{header}{display header if TRUE}

\item{first_row}{Specifies that the first column conditional formatting should be
applied. Details for this and other conditional formatting options can be found
at http://officeopenxml.com/WPtblLook.php.}

\item{first_column}{Specifies that the first column conditional formatting should
be applied.}

\item{last_row}{Specifies that the first column conditional formatting should be applied.}

\item{last_column}{Specifies that the first column conditional formatting should be
applied.}

\item{no_hband}{Specifies that the first column conditional formatting should be applied.}

\item{no_vband}{Specifies that the first column conditional formatting should be applied.}
}
\description{
Create a representation of a table
}
\examples{
block_table(x = mtcars)
}
\seealso{
Other block functions for reporting: 
\code{\link{block_caption}()},
\code{\link{block_list}()},
\code{\link{block_section}()},
\code{\link{block_toc}()},
\code{\link{fpar}()},
\code{\link{plot_instr}()},
\code{\link{unordered_list}()}
}
\concept{block functions for reporting}
