% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_add.R
\name{body_add_caption}
\alias{body_add_caption}
\title{add caption into an rdocx object}
\usage{
body_add_caption(x, value, pos = "after")
}
\arguments{
\item{x}{a docx device}

\item{value}{a \link{block_caption} object.}

\item{pos}{where to add the new element relative to the cursor,
one of "after", "before", "on".}
}
\description{
add a \code{caption} into an rdocx object. A Caption
is created with function \link{block_caption}.
}
\examples{
library(magrittr)
run_num <- run_autonum(seq_id = "tab", pre_label = "tab. ")
caption <- block_caption("iris table",
  style = "Normal", id = "iris_table",
  autonum = run_num )

doc <- read_docx() \%>\%
  body_add_par("A title", style = "heading 1") \%>\%
  body_add_par("Hello world!", style = "Normal") \%>\%
  body_add_caption(caption) \%>\%
  body_add_table(iris, style = "table_template")

print(doc, target = tempfile(fileext = ".docx") )
}
