% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_add.R
\name{body_add_img}
\alias{body_add_img}
\title{add image}
\usage{
body_add_img(x, src, style = "Normal", width, height, pos = "after")
}
\arguments{
\item{x}{a docx device}

\item{src}{image filename}

\item{style}{paragraph style}

\item{width}{height in inches}

\item{height}{height in inches}

\item{pos}{where to add the new element relative to the cursor,
one of "after", "before", "on".}
}
\description{
add an image into a docx object
}
\examples{
doc <- read_docx()

img.file <- file.path( Sys.getenv("R_HOME"), "doc", "html", "logo.jpg" )
if( file.exists(img.file) ){
  doc <- body_add_img(x = doc, src = img.file, height = 1.06, width = 1.39 )
}
if( require("ionicons") ){
  calendar_src = as_png(name = "calendar", fill = "#FFE64D", width = 144, height = 144)
  doc <- body_add_img(x = doc, src = calendar_src, height = 2, width = 2 )
}

print(doc, target = "body_add_img.docx" )
}
