% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_add.R
\name{body_add_img}
\alias{body_add_img}
\title{add image}
\usage{
body_add_img(x, src, style = NULL, width, height, pos = "after")
}
\arguments{
\item{x}{an rdocx object}

\item{src}{image filename, the basename of the file must not contain any blank.}

\item{style}{paragraph style}

\item{width}{height in inches}

\item{height}{height in inches}

\item{pos}{where to add the new element relative to the cursor,
one of "after", "before", "on".}
}
\description{
add an image into an rdocx object.
}
\examples{
doc <- read_docx()

img.file <- file.path( R.home("doc"), "html", "logo.jpg" )
if( file.exists(img.file) ){
  doc <- body_add_img(x = doc, src = img.file, height = 1.06, width = 1.39 )
}

print(doc, target = tempfile(fileext = ".docx"))
}
