% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_add.R
\name{body_add_par}
\alias{body_add_par}
\title{add paragraph of text}
\usage{
body_add_par(x, value, style = NULL, pos = "after")
}
\arguments{
\item{x}{a docx device}

\item{value}{a character}

\item{style}{paragraph style}

\item{pos}{where to add the new element relative to the cursor,
one of "after", "before", "on".}
}
\description{
add a paragraph of text into an rdocx object
}
\examples{
library(magrittr)

doc <- read_docx() \%>\%
  body_add_par("A title", style = "heading 1") \%>\%
  body_add_par("Hello world!", style = "Normal") \%>\%
  body_add_par("centered text", style = "centered")

print(doc, target = "body_add_par.docx" )
}
