% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_add.R
\name{body_add_table}
\alias{body_add_table}
\title{add table}
\usage{
body_add_table(x, value, style, pos = "after", first_row = TRUE,
  first_column = FALSE, last_row = FALSE, last_column = FALSE,
  no_hband = FALSE, no_vband = TRUE)
}
\arguments{
\item{x}{a docx device}

\item{value}{a data.frame}

\item{style}{table style}

\item{pos}{where to add the new element relative to the cursor,
one of "after", "before", "on".}

\item{first_row, last_row, first_column, last_column, no_hband, no_vband}{logical for Word table options}
}
\description{
add a table into a docx object
}
\examples{
library(magrittr)

doc <- read_docx() \%>\%
  body_add_table(iris, style = "table_template")
print(doc, target = "body_add_table.docx" )
}
