% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_section.R
\name{body_end_section_columns}
\alias{body_end_section_columns}
\title{Add multi columns section}
\usage{
body_end_section_columns(x, widths = c(2.5, 2.5), space = 0.25, sep = FALSE)
}
\arguments{
\item{x}{an rdocx object}

\item{widths}{columns widths in inches. If 3 values, 3 columns
will be produced.}

\item{space}{space in inches between columns.}

\item{sep}{if TRUE a line is separating columns.}
}
\description{
A section with multiple columns is added to the
document.

You may prefer to use \code{\link[=body_end_block_section]{body_end_block_section()}} that is
more flexible.
}
\examples{
str1 <- "Lorem ipsum dolor sit amet, consectetur adipiscing elit."
str1 <- rep(str1, 5)
str1 <- paste(str1, collapse = " ")

doc_1 <- read_docx()
doc_1 <- body_add_par(doc_1, value = str1, style = "Normal")
doc_1 <- body_add_par(doc_1, value = str1, style = "Normal")
doc_1 <- body_end_section_columns(doc_1)
doc_1 <- body_add_par(doc_1, value = str1, style = "Normal")
print(doc_1, target = tempfile(fileext = ".docx"))
}
\seealso{
Other functions for Word sections: 
\code{\link{body_end_block_section}()},
\code{\link{body_end_section_columns_landscape}()},
\code{\link{body_end_section_continuous}()},
\code{\link{body_end_section_landscape}()},
\code{\link{body_end_section_portrait}()},
\code{\link{body_set_default_section}()}
}
\concept{functions for Word sections}
