% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_section.R
\name{body_end_section_columns_landscape}
\alias{body_end_section_columns_landscape}
\title{add multi columns section within landscape orientation}
\usage{
body_end_section_columns_landscape(
  x,
  widths = c(2.5, 2.5),
  space = 0.25,
  sep = FALSE,
  w = 21/2.54,
  h = 29.7/2.54
)
}
\arguments{
\item{x}{an rdocx object}

\item{widths}{columns widths in inches. If 3 values, 3 columns
will be produced.}

\item{space}{space in inches between columns.}

\item{sep}{if TRUE a line is separating columns.}

\item{w, h}{page width, page height (in inches)}
}
\description{
A landscape section with multiple columns is added to the document.
}
\examples{
str1 <- "Lorem ipsum dolor sit amet, consectetur adipiscing elit."
str1 <- rep(str1, 5)
str1 <- paste(str1, collapse = " ")

doc_1 <- read_docx()
doc_1 <- body_add_par(doc_1, value = str1, style = "Normal")
doc_1 <- slip_in_column_break(doc_1, pos = "after")
doc_1 <- body_add_par(doc_1, value = str1, style = "Normal")
doc_1 <- body_end_section_columns_landscape(doc_1, widths = c(6, 2))
doc_1 <- body_add_par(doc_1, value = str1, style = "Normal")
print(doc_1, target = tempfile(fileext = ".docx"))
}
\seealso{
Other functions for Word sections: 
\code{\link{body_end_block_section}()},
\code{\link{body_end_section_columns}()},
\code{\link{body_end_section_continuous}()},
\code{\link{body_end_section_landscape}()},
\code{\link{body_end_section_portrait}()}
}
\concept{functions for Word sections}
