% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_section.R
\name{body_end_section_portrait}
\alias{body_end_section_portrait}
\title{add portrait section}
\usage{
body_end_section_portrait(x, w = 21/2.54, h = 29.7/2.54)
}
\arguments{
\item{x}{an rdocx object}

\item{w, h}{page width, page height (in inches)}
}
\description{
A section with portrait orientation is added to the document.
}
\examples{
str1 <- "Lorem ipsum dolor sit amet, consectetur adipiscing elit."
str1 <- rep(str1, 5)
str1 <- paste(str1, collapse = " ")

doc_1 <- read_docx()
doc_1 <- body_add_par(doc_1, value = str1, style = "Normal")
doc_1 <- body_end_section_portrait(doc_1)
doc_1 <- body_add_par(doc_1, value = str1, style = "Normal")
print(doc_1, target = tempfile(fileext = ".docx"))
}
\seealso{
Other functions for Word sections: 
\code{\link{body_end_block_section}()},
\code{\link{body_end_section_columns_landscape}()},
\code{\link{body_end_section_columns}()},
\code{\link{body_end_section_continuous}()},
\code{\link{body_end_section_landscape}()}
}
\concept{functions for Word sections}
