% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pptx.R
\name{color_scheme}
\alias{color_scheme}
\title{color scheme}
\usage{
color_scheme(x)
}
\arguments{
\item{x}{pptx object}
}
\description{
get master layout color scheme into a data.frame.
}
\examples{
x <- read_pptx()
color_scheme ( x = x )
}
