% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_docx.R
\name{doc_properties}
\alias{doc_properties}
\title{read document properties}
\usage{
doc_properties(x)
}
\arguments{
\item{x}{an \code{rdocx} or \code{rpptx} object}
}
\description{
read Word or PowerPoint document properties
and get results in a tidy data.frame.
}
\examples{
library(magrittr)
read_docx() \%>\% doc_properties()
}
