% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_docx.R
\name{doc_properties}
\alias{doc_properties}
\title{read document properties}
\usage{
doc_properties(x)
}
\arguments{
\item{x}{an \code{rdocx} or \code{rpptx} object}
}
\description{
read Word or PowerPoint document properties
and get results in a data.frame.
}
\examples{
x <- read_docx()
doc_properties(x)
}
\seealso{
Other functions for Word document informations: 
\code{\link{docx_bookmarks}()},
\code{\link{docx_dim}()},
\code{\link{length.rdocx}()},
\code{\link{set_doc_properties}()},
\code{\link{styles_info}()}
}
\concept{functions for Word document informations}
