% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_docx.R
\name{doc_properties}
\alias{doc_properties}
\title{Read document properties}
\usage{
doc_properties(x)
}
\arguments{
\item{x}{an \code{rdocx} or \code{rpptx} object}
}
\value{
a data.frame
}
\description{
Read Word or PowerPoint document properties
and get results in a data.frame.
}
\examples{
x <- read_docx()
doc_properties(x)
}
\seealso{
Other functions for Word document informations: 
\code{\link{docx_bookmarks}()},
\code{\link{docx_dim}()},
\code{\link{length.rdocx}()},
\code{\link{set_doc_properties}()},
\code{\link{styles_info}()}

Other functions for reading presentation information: 
\code{\link{annotate_base}()},
\code{\link{color_scheme}()},
\code{\link{layout_properties}()},
\code{\link{layout_summary}()},
\code{\link{length.rpptx}()},
\code{\link{plot_layout_properties}()},
\code{\link{slide_size}()},
\code{\link{slide_summary}()}
}
\concept{functions for Word document informations}
\concept{functions for reading presentation information}
