% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_docx.R
\name{docx_bookmarks}
\alias{docx_bookmarks}
\title{List Word bookmarks}
\usage{
docx_bookmarks(x)
}
\arguments{
\item{x}{an \code{rdocx} object}
}
\description{
List bookmarks id that can be found in an \code{rdocx}
object.
}
\examples{
library(magrittr)

doc <- read_docx() \%>\%
  body_add_par("centered text", style = "centered") \%>\%
  body_bookmark("text_to_replace") \%>\% body_add_par("centered text", style = "centered") \%>\%
  body_bookmark("text_to_replace2")

docx_bookmarks(doc)

docx_bookmarks(read_docx())
}
