% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_utils_funs.R
\name{docx_dim}
\alias{docx_dim}
\title{'Word' page layout}
\usage{
docx_dim(x)
}
\arguments{
\item{x}{an \code{rdocx} object}
}
\description{
Get page width, page height and margins (in inches). The return values
are those corresponding to the section where the cursor is.
}
\examples{
docx_dim(read_docx())
}
\seealso{
Other functions for Word document informations: 
\code{\link{doc_properties}()},
\code{\link{docx_bookmarks}()},
\code{\link{length.rdocx}()},
\code{\link{set_doc_properties}()},
\code{\link{styles_info}()}
}
\concept{functions for Word document informations}
