% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{docx_reference_img}
\alias{docx_reference_img}
\title{add images into an rdocx object}
\usage{
docx_reference_img(x, src)
}
\arguments{
\item{x}{an rdocx object}

\item{src}{a vector of character containing image filenames.}
}
\description{
reference images into a Word document.
This function is now useless as the processing of images
is automated when using \code{\link[=print.rdocx]{print.rdocx()}}.
}
\keyword{internal}
