% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-proc.R
\name{docx_reference_img}
\alias{docx_reference_img}
\title{add images into an rdocx object}
\usage{
docx_reference_img(x, src)
}
\arguments{
\item{x}{an rdocx object}

\item{src}{a vector of character containing image filenames.}
}
\description{
reference images into a Word document.
This function is now useless as the processing of images
is automated when using \code{\link[=print.rdocx]{print.rdocx()}}.
}
\seealso{
Other functions for officer extensions: 
\code{\link{fortify_location}()},
\code{\link{get_reference_value}()},
\code{\link{opts_current_table}()},
\code{\link{to_html}()},
\code{\link{to_pml}()},
\code{\link{to_wml}()},
\code{\link{wml_link_images}()}
}
\concept{functions for officer extensions}
\keyword{internal}
