% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_run_objects.R
\name{external_img}
\alias{external_img}
\title{external image}
\usage{
external_img(src, width = 0.5, height = 0.2)
}
\arguments{
\item{src}{image file path}

\item{width}{height in inches}

\item{height}{height in inches}
}
\description{
This function is used to insert images in 'PowerPoint'
slides.
}
\examples{
img.file <- file.path( R.home("doc"), "html", "logo.jpg" )

doc <- read_pptx()
doc <- add_slide(doc)
doc <- ph_with(x = doc,
  value = external_img(img.file, width = 1.39, height = 1.06),
  location = ph_location_type(type = "body"),
  use_loc_size = FALSE )
print(doc, target = tempfile(fileext = ".pptx"))
}
\seealso{
\link{ph_with}
}
