% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cot_objects.R
\name{external_img}
\alias{external_img}
\alias{dim.external_img}
\alias{as.data.frame.external_img}
\alias{format.external_img}
\title{external image}
\usage{
external_img(src, width = 0.5, height = 0.2)

\method{dim}{external_img}(x)

\method{as.data.frame}{external_img}(x, ...)

\method{format}{external_img}(x, type = "console", ...)
}
\arguments{
\item{src}{image file path}

\item{width}{height in inches}

\item{height}{height in inches}

\item{x}{\code{external_img} object}

\item{...}{unused}

\item{type}{output format}
}
\description{
This function is used to insert images into flextable with function \code{display}
}
\examples{
# external_img("example.png")
}
