% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_border.R
\name{fp_border}
\alias{fp_border}
\alias{update.fp_border}
\alias{format.fp_border}
\title{border properties object}
\usage{
fp_border(color = "black", style = "solid", width = 1)

\method{update}{fp_border}(object, color, style, width, ...)

\method{format}{fp_border}(x, type = "pml", ...)
}
\arguments{
\item{color}{border color - single character value (e.g. "#000000" or "black")}

\item{style}{border style - single character value : "none" or "solid" or "dotted" or "dashed"}

\item{width}{border width - an integer value : 0>= value}

\item{...}{further arguments - not used}

\item{x, object}{\code{fp_border} object}

\item{type}{output type - only 'pml' currently supported.}
}
\description{
create a border properties object.
}
\examples{
fp_border()
fp_border(color="orange", style="solid", width=1)
fp_border(color="gray", style="dotted", width=1)

# modify object ------
border <- fp_border()
update(border, style="dotted", width=3)
}
