% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpar.R
\name{fpar}
\alias{fpar}
\alias{update.fpar}
\alias{fortify_fpar}
\alias{as.data.frame.fpar}
\alias{format.fpar}
\alias{print.fpar}
\title{concatenate formatted text}
\usage{
fpar(...)

\method{update}{fpar}(object, fp_p = NULL, fp_t = NULL, ...)

fortify_fpar(x)

\method{as.data.frame}{fpar}(x, ...)

\method{format}{fpar}(x, type = "pml", ...)

\method{print}{fpar}(x, ...)
}
\arguments{
\item{...}{unused}

\item{fp_p}{paragraph formatting properties}

\item{fp_t}{default text formatting properties}

\item{x, object}{fpar object}

\item{type}{a string value ("pml", "wml" or "html").}
}
\description{
Create a paragraph representation by concatenating
formatted text or images. Modify default text and paragraph
formatting properties with update.
}
\details{
\code{fortify_fpar}, \code{as.data.frame} are used internally and
are not supposed to be used by end user.
}
\examples{
fpar(ftext("hello", shortcuts$fp_bold()))
}
