% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled.R
\name{from_labelled}
\alias{from_labelled}
\alias{from_labelled.data.frame}
\alias{from_labelled.labelled}
\title{Convert from labelled data}
\usage{
from_labelled(x)

\method{from_labelled}{labelled}(x)

\method{from_labelled}{data.frame}(x)
}
\arguments{
\item{x}{A \code{\link[haven]{labelled}} vector or \code{data.frame} containing it.}
}
\description{
When reading SPSS files with \code{\link{read_data}}, the output contains vectors
of class \code{labelled}. This function extracts the labels and attaches them
as an attribute (by the same name) to the data, and converts any \code{labelled}
variables into factors.
}
\note{
\code{\link[data.table]{data.table}} input returns a copy of the \code{data.table}.
}
\examples{
vals <- c("Agree", "Neutral", "Disagree", "Don't know")
var <- haven::labelled(c(1, 5), labels = setNames(1:4, vals))

# Works with just the labelled vector
from_labelled(var)

# And a data.frame containing labelled vectors
from_labelled(as.data.frame(var))$var
}
\author{
Kristian D. Olsen
}

