% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_run_objects.R
\name{ftext}
\alias{ftext}
\title{formatted text}
\usage{
ftext(text, prop)
}
\arguments{
\item{text}{text value}

\item{prop}{formatting text properties}
}
\description{
Format a chunk of text with text formatting properties.
}
\examples{
ftext("hello", fp_text())
}
