% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cot_objects.R
\name{ftext}
\alias{ftext}
\alias{format.ftext}
\alias{print.ftext}
\title{formatted text}
\usage{
ftext(text, prop)

\method{format}{ftext}(x, type = "console", ...)

\method{print}{ftext}(x, ...)
}
\arguments{
\item{text}{text value}

\item{prop}{formatting text properties}

\item{x}{\code{ftext} object}

\item{type}{output format, one of wml, pml, html, console, text.}

\item{...}{unused}
}
\description{
Format a chunk of text with text formatting properties.
}
\examples{
ftext("hello", fp_text())
}
