% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx_layout_helper.R
\name{get_layout}
\alias{get_layout}
\title{Layout selection helper}
\usage{
get_layout(
  x,
  layout = NULL,
  master = NULL,
  layout_by_id = TRUE,
  get_first = FALSE
)
}
\arguments{
\item{x}{An \code{rpptx} object.}

\item{layout}{Layout name or index. Index refers to the row index of the \code{\link[=layout_summary]{layout_summary()}}
output.}

\item{master}{Name of master. Only required if layout name is not unique across masters.}

\item{layout_by_id}{Allow layout index instead of name? (default is \code{TRUE})}

\item{get_first}{If layout exists in multiple master, return first occurence (default \code{FALSE}).}
}
\value{
A \verb{<layout_info>} object, i.e. a list with the entries \code{index}, \code{layout_name},
\code{layout_file}, \code{master_name}, \code{master_file}, and \code{slide_layout}.
}
\description{
Select a layout by name or index. The master name is inferred and only required
for disambiguation in case the layout name is not unique across masters.
}
\keyword{internal}
