% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_595_plot_layout_properties_improved.R
\name{get_layout}
\alias{get_layout}
\title{Layout selection helper}
\usage{
get_layout(x, layout, master = NULL)
}
\arguments{
\item{x}{An \code{rpptx} object.}

\item{layout}{Layout name or index, see \link{layout_summary}.}

\item{master}{Name of master. Only required if layout name is not unique across masters.}
}
\value{
A list with layout's index, name, and master.
}
\description{
Select a layout by name or index. Master name is inferred and only required
for disambiguation in case of several masters.
}
\examples{
x <- read_pptx()
get_layout(x, "Title Slide")

}
\keyword{internal}
