% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pptx.R
\name{layout_properties}
\alias{layout_properties}
\title{slide layout properties}
\usage{
layout_properties(x, layout = NULL, master = NULL)
}
\arguments{
\item{x}{rpptx object}

\item{layout}{slide layout name to use}

\item{master}{master layout name where \code{layout} is located}
}
\description{
get information about a particular slide layout
into a data.frame.
}
\examples{
x <- read_pptx()
layout_properties ( x = x, layout = "Title Slide", master = "Office Theme" )
layout_properties ( x = x, master = "Office Theme" )
layout_properties ( x = x, layout = "Two Content" )
layout_properties ( x = x )
}
