% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pptx.R
\name{layout_summary}
\alias{layout_summary}
\title{presentation layouts summary}
\usage{
layout_summary(x)
}
\arguments{
\item{x}{rpptx object}
}
\description{
get informations about slide layouts and
master layouts into a data.frame.
}
\examples{
my_pres <- read_pptx()
layout_summary ( x = my_pres )
}
