% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx_informations.R
\name{length.rpptx}
\alias{length.rpptx}
\title{Number of slides}
\usage{
\method{length}{rpptx}(x)
}
\arguments{
\item{x}{an rpptx object}
}
\description{
Function \code{length} will return the number of slides.
}
\examples{
my_pres <- read_pptx()
my_pres <- add_slide(my_pres, "Title and Content")
my_pres <- add_slide(my_pres, "Title and Content")
length(my_pres)
}
\seealso{
Other functions for reading presentation information:
\code{\link{add_annotated_layouts}()},
\code{\link{annotate_base}()},
\code{\link{color_scheme}()},
\code{\link{doc_properties}()},
\code{\link{layout_properties}()},
\code{\link{layout_summary}()},
\code{\link{phs_annotate}()},
\code{\link{plot_layout_properties}()},
\code{\link{slide_size}()},
\code{\link{slide_summary}()}
}
\concept{ppt_info}
