% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_ph_with_methods.R
\name{location_eval}
\alias{location_eval}
\title{eval a location on the current slide}
\usage{
location_eval(location, x)
}
\arguments{
\item{location}{a location for a placeholder. It has to be a
quosure.}

\item{x}{an rpptx object}
}
\description{
Eval a shape location against the current slide.
This function is to be used to add custom openxml code.
}
\examples{
library(rlang)
doc <- read_pptx()
doc <- add_slide(doc, layout = "Title and Content",
  master = "Office Theme")
location_eval(quo(ph_location_fullsize()), doc)
}
\seealso{
\code{\link{ph_location}}, \code{\link{ph_with}}
}
