% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pptx.R
\name{move_slide}
\alias{move_slide}
\title{move a slide}
\usage{
move_slide(x, index, to)
}
\arguments{
\item{x}{rpptx object}

\item{index}{slide index, default to current slide position.}

\item{to}{new slide index.}
}
\description{
move a slide in a pptx presentation
}
\note{
cursor is set on the last slide.
}
\examples{
x <- read_pptx()
x <- add_slide(x, layout = "Title and Content",
  master = "Office Theme")
x <- ph_with_text(x, type = "body", str = "Hello world 1")
x <- add_slide(x, layout = "Title and Content",
  master = "Office Theme")
x <- ph_with_text(x, type = "body", str = "Hello world 2")
x <- move_slide(x, index = 1, to = 2)
}
