% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{officer-defunct}
\alias{officer-defunct}
\alias{ph_with_gg_at}
\alias{ph_with_table_at}
\alias{ph_with_text}
\title{Defunct Functions in Package officer}
\usage{
ph_with_gg_at(...)

ph_with_table_at(...)

ph_with_text(...)
}
\arguments{
\item{...}{unused arguments}
}
\description{
Defunct Functions in Package officer
}
\details{
\code{ph_with()} is replaced by \code{ph_with.gg}.

\code{ph_with_table_at()} is replaced by \code{ph_with.data.frame}.

\code{ph_with_text()} is replaced by \code{ph_with.character}.
}
