% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{officer-defunct}
\alias{officer-defunct}
\alias{ph_from_xml}
\alias{ph_from_xml_at}
\alias{ph_with_table}
\alias{ph_with_img}
\alias{ph_with_gg}
\alias{ph_with_ul}
\alias{ph_with_table_at}
\alias{ph_with_fpars_at}
\title{Defunct Functions in Package officer}
\usage{
ph_from_xml(...)

ph_from_xml_at(...)

ph_with_table(...)

ph_with_img(...)

ph_with_gg(...)

ph_with_ul(...)

ph_with_table_at(...)

ph_with_fpars_at(...)
}
\arguments{
\item{...}{unused arguments}
}
\description{
Defunct Functions in Package officer
}
\details{
\code{ph_from_xml()} is replaced by \code{ph_with.xml_document}.

\code{ph_from_xml_at()} is replaced by \code{ph_with.xml_document}.

\code{ph_with_table()} is replaced by \code{ph_with.xml_document}.

\code{ph_with_img()} is replaced by \code{ph_with.xml_document}.

\code{ph_with_gg()} is replaced by \code{ph_with.xml_document}.

\code{ph_with_ul()} is replaced by \code{ph_with.xml_document}.

\code{ph_with_table_at()} is replaced by \code{ph_with.xml_document}.

\code{ph_with_fpars_at()} is replaced by \code{ph_with.xml_document}.
}
