% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{officer-defunct}
\alias{officer-defunct}
\alias{ph_add_text}
\alias{ph_add_par}
\alias{ph_add_fpar}
\alias{slip_in_seqfield}
\alias{slip_in_img}
\title{Defunct Functions in Package officer}
\usage{
ph_add_text(...)

ph_add_par(...)

ph_add_fpar(...)

slip_in_seqfield(...)

slip_in_img(...)
}
\arguments{
\item{...}{unused arguments}
}
\description{
Defunct Functions in Package officer
}
\details{
\code{ph_add_text()} is replaced by \code{ph_with(value = fpar(...))}.

\code{ph_add_par()} is replaced by \code{ph_with(value = fpar(...))}.

\code{ph_add_fpar()} is replaced by \code{ph_with(value = fpar(...))}.

\code{slip_in_seqfield()} is replaced by \code{run_word_field()}.

\code{slip_in_img()} is replaced by \code{external_img()}.
}
