% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{officer-defunct}
\alias{officer-defunct}
\alias{ph_with_gg_at}
\alias{ph_with_table_at}
\alias{ph_with_img_at}
\alias{ph_with_img}
\alias{ph_with_text}
\alias{ph_empty_at}
\alias{ph_empty}
\title{Defunct Functions in Package officer}
\usage{
ph_with_gg_at(...)

ph_with_table_at(...)

ph_with_img_at(...)

ph_with_img(...)

ph_with_text(...)

ph_empty_at(...)

ph_empty(...)
}
\arguments{
\item{...}{unused arguments}
}
\description{
Defunct Functions in Package officer
}
\details{
\code{ph_with()} is replaced by \code{ph_with.gg}.

\code{ph_with_table_at()} is replaced by \code{ph_with.data.frame}.

\code{ph_with_img_at()} is replaced by \code{ph_with.external_img}.

\code{ph_with_img()} is replaced by \code{ph_with.external_img}.

\code{ph_with_text()} is replaced by \code{ph_with.character}.

\code{ph_empty_at()} is replaced by \code{ph_with.empty_content}.

\code{ph_empty()} is replaced by \code{ph_with.empty_content}.
}
