% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{officer-defunct}
\alias{officer-defunct}
\alias{slip_in_seqfield}
\alias{slip_in_column_break}
\alias{slip_in_xml}
\alias{slip_in_text}
\alias{slip_in_footnote}
\title{Defunct Functions in Package officer}
\usage{
slip_in_seqfield(...)

slip_in_column_break(...)

slip_in_xml(...)

slip_in_text(...)

slip_in_footnote(...)
}
\arguments{
\item{...}{unused arguments}
}
\description{
Defunct Functions in Package officer
}
\details{
\code{slip_in_seqfield()} is replaced by \code{run_word_field()}.

\code{slip_in_column_break()} is replaced by \code{run_columnbreak()}.

\code{slip_in_xml()} is replaced by \code{fpar()}.

\code{slip_in_text()} is replaced by \code{fpar()}.

\code{slip_in_footnote()} is replaced by \code{run_footnote()}.
}
