% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.R
\docType{package}
\name{officer}
\alias{officer}
\alias{officer-package}
\title{officer: Manipulate Microsoft Word and PowerPoint Documents}
\description{
The officer package facilitates access to and manipulation of
'Microsoft Word' and 'Microsoft PowerPoint' documents from R.
}
\details{
\itemize{
\item read Word and PowerPoint files into data objects
\item add/edit/remove image, table and text content from documents and slides
\item write updated content back to Word and PowerPoint files
}

To learn more about officer, start with the vignettes:
`browseVignettes(package = "officer")`
}
\seealso{
Useful links:
\itemize{
  \item \url{https://davidgohel.github.io/officer}
  \item Report bugs at \url{https://github.com/davidgohel/officer/issues}
}

}
\author{
\strong{Maintainer}: David Gohel \email{david.gohel@ardata.fr}

Other contributors:
\itemize{
  \item Frank Hangler \email{frank@plotandscatter.com} (function body_replace_all_text) [contributor]
  \item Liz Sander \email{lsander@civisanalytics.com} (several documentation fixes) [contributor]
  \item Jon Calder \email{jonmcalder@gmail.com} (update vignettes) [contributor]
}

}
