% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.R
\docType{package}
\name{officer}
\alias{officer}
\alias{officer-package}
\title{officer: Manipulate Microsoft Word and PowerPoint Documents}
\description{
The officer package facilitates access to and manipulation of
'Microsoft Word' and 'Microsoft PowerPoint' documents from R.
}
\details{
Examples of manipulations are:

\itemize{
\item read Word and PowerPoint files into data objects
\item add/edit/remove image, table and text content from documents and slides
\item write updated content back to Word and PowerPoint files
}

To learn more about officer, start with the vignettes:
\code{browseVignettes(package = "officer")}
}
\seealso{
\url{https://davidgohel.github.io/officer/}
}
\author{
\strong{Maintainer}: David Gohel \email{david.gohel@ardata.fr}

Other contributors:
\itemize{
  \item Frank Hangler \email{frank@plotandscatter.com} (function body_replace_all_text) [contributor]
  \item Liz Sander \email{lsander@civisanalytics.com} (several documentation fixes) [contributor]
  \item Anton Victorson \email{anton@victorson.se} (fixes xml structures) [contributor]
  \item Jon Calder \email{jonmcalder@gmail.com} (update vignettes) [contributor]
  \item John Harrold \email{john.m.harrold@gmail.com} (fuction annotate_base) [contributor]
}

}
