% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pptx.R
\name{on_slide}
\alias{on_slide}
\title{change current slide}
\usage{
on_slide(x, index)
}
\arguments{
\item{x}{rpptx object}

\item{index}{slide index}
}
\description{
change current slide index of an rpptx object.
}
\examples{
doc <- read_pptx()
doc <- add_slide(doc, layout = "Title and Content", master = "Office Theme")
doc <- add_slide(doc, layout = "Title and Content", master = "Office Theme")
doc <- add_slide(doc, layout = "Title and Content", master = "Office Theme")
doc <- on_slide( doc, index = 1)
doc <- ph_with_text(x = doc, type = "title", str = "First title")
doc <- on_slide( doc, index = 3)
doc <- ph_with_text(x = doc, type = "title", str = "Third title")

print(doc, target = "on_slide.pptx" )
}
