% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_run_objects.R
\name{page_size}
\alias{page_size}
\title{page size object}
\usage{
page_size(width = 21/2.54, height = 29.7/2.54, orient = "portrait")
}
\arguments{
\item{width, height}{page width, page height (in inches).}

\item{orient}{page orientation, either 'landscape', either 'portrait'.}
}
\description{
The function creates a representation of the dimensions of a page. The
dimensions are defined by length, width and orientation. If the orientation is
in landscape mode then the length becomes the width and the width becomes the length.
}
\examples{
page_size(orient = "landscape")
}
\seealso{
Other functions for section definition: 
\code{\link{page_mar}()},
\code{\link{prop_section}()},
\code{\link{section_columns}()}
}
\concept{functions for section definition}
