% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{ph_add_text}
\alias{ph_add_text}
\title{append text}
\usage{
ph_add_text(
  x,
  str,
  type = "body",
  id = 1,
  id_chr = NULL,
  ph_label = NULL,
  style = fp_text(font.size = 0),
  pos = "after",
  href = NULL,
  slide_index = NULL
)
}
\arguments{
\item{x}{an rpptx object}

\item{str}{text to add}

\item{type}{placeholder type}

\item{id}{placeholder index (integer) for a duplicated type. This is to be used when a placeholder
type is not unique in the layout of the current slide, e.g. two placeholders with type 'body'. To
add onto the first, use \code{id = 1} and \code{id = 2} for the second one.
Values can be read from \code{\link{slide_summary}}.}

\item{id_chr}{deprecated.}

\item{ph_label}{label associated to the placeholder. Use column
\code{ph_label} of result returned by \code{\link{slide_summary}}.}

\item{style}{text style, a \code{\link{fp_text}} object}

\item{pos}{where to add the new element relative to the cursor,
"after" or "before".}

\item{href}{hyperlink to reach when clicking the text}

\item{slide_index}{slide index to reach when clicking the text.
It will be ignored if \code{href} is not NULL.}
}
\description{
append text in a placeholder.
The function let you add text to an existing
content in an exiisting shape, existing text will be preserved.
}
\note{
This function will be deprecated in a next release because it is not
efficient and make users write complex code. Use instead \code{\link[=fpar]{fpar()}} to build
formatted paragraphs.
}
\section{Usage}{

If your goal is to add formatted text in a new shape, use \code{\link{ph_with}}
with a \code{\link{block_list}} instead of this function.
}

