% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_ph_with.R
\name{ph_empty}
\alias{ph_empty}
\title{add a new empty shape}
\usage{
ph_empty(x, type = "title", index = 1)
}
\arguments{
\item{x}{a pptx device}

\item{type}{placeholder type}

\item{index}{placeholder index (integer). This is to be used when a placeholder type
is not unique in the current slide, e.g. two placeholders with type 'body'.}
}
\description{
add a new empty shape in the current slide.
}
\examples{
fileout <- tempfile(fileext = ".pptx")
doc <- read_pptx()
doc <- add_slide(doc, layout = "Title and Content", master = "Office Theme")
doc <- ph_empty(x = doc, type = "title")

print(doc, target = fileout )
}
