% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_ph_with.R
\name{ph_from_xml}
\alias{ph_from_xml}
\title{add an xml string as new shape}
\usage{
ph_from_xml(x, value, type = "body", index = 1)
}
\arguments{
\item{x}{a pptx device}

\item{value}{a character}

\item{type}{placeholder type}

\item{index}{placeholder index (integer). This is to be used when a placeholder type
is not unique in the current slide, e.g. two placeholders with type 'body'.}
}
\description{
Add an xml string as new shape in the current slide. This function
is to be used to add custom openxml code.
}
