% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_location.R
\name{ph_location}
\alias{ph_location}
\title{create a location for a placeholder}
\usage{
ph_location(left = 1, top = 1, width = 4, height = 3, label = "",
  ph = "<p:ph/>", bg = NULL, rotation = NULL, ...)
}
\arguments{
\item{left, top, width, height}{place holder coordinates
in inches.}

\item{label}{a label for the placeholder. See section details.}

\item{ph}{openxml string value for tag. This is not meant to be used
by end users (unless they understand openxml placeholder and
\code{p:ph} tag).}

\item{bg}{background color}

\item{rotation}{rotation angle}

\item{...}{unused arguments}
}
\description{
The function will return a list that complies with
expected format for argument \code{location} of functions \code{ph_with_*}
and \code{ph_with} methods.
}
\details{
The location of the bounding box associated to a placeholder
within a presentation slide is specified with the left top coordinate,
the width and the height. These are defined in inches:

\describe{
\item{left}{left coordinate of the bounding box}
\item{top}{top coordinate of the bounding box}
\item{width}{width of the bounding box}
\item{height}{height of the bounding box}
}

In addition to these attributes, there is attribute \code{ph_label}
associated with the shape (shapes, text boxes, images and other objects
will be identified with that label in the Selection Pane of PowerPoint).
This label can then be reused by other functions such as \code{ph_add_fpar},
or \code{ph_add_text}.
}
\examples{
library(magrittr)
read_pptx() \%>\%
  add_slide(layout = "Title and Content", master = "Office Theme") \%>\%
  ph_with("Hello world",
    location = ph_location(width = 4, height = 3, label = "hello") ) \%>\%
  print(target = tempfile(fileext = ".pptx") )
}
\seealso{
Other functions for placeholder location: \code{\link{ph_location_fullsize}},
  \code{\link{ph_location_label}},
  \code{\link{ph_location_left}},
  \code{\link{ph_location_right}},
  \code{\link{ph_location_type}}
}
\concept{functions for placeholder location}
