% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_location.R
\name{ph_location_fullsize}
\alias{ph_location_fullsize}
\title{location of a full size element}
\usage{
ph_location_fullsize(x, label = "")
}
\arguments{
\item{x}{an rpptx object}

\item{label}{a label to associate with the placeholder.}
}
\description{
The function will return the location corresponding
to a full size display.
}
\examples{
x <- read_pptx()
ph_location_fullsize(x)
}
\seealso{
Other functions for placeholder location: \code{\link{ph_location_label}},
  \code{\link{ph_location_left}},
  \code{\link{ph_location_right}},
  \code{\link{ph_location_type}}, \code{\link{ph_location}}
}
\concept{functions for placeholder location}
