% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_location.R
\name{ph_location_fullsize}
\alias{ph_location_fullsize}
\title{location of a full size element}
\usage{
ph_location_fullsize(label = "", ...)
}
\arguments{
\item{label}{a label to associate with the placeholder.}

\item{...}{unused arguments}
}
\description{
The function will return the location corresponding
to a full size display.
}
\examples{
library(magrittr)
my_pres <- read_pptx() \%>\%
  add_slide(layout = "Title and Content", master = "Office Theme") \%>\%
  ph_with("Hello world", location = ph_location_fullsize() ) \%>\%
  print(target = tempfile(fileext = ".pptx") )
}
\seealso{
Other functions for placeholder location: \code{\link{ph_location_label}},
  \code{\link{ph_location_left}},
  \code{\link{ph_location_right}},
  \code{\link{ph_location_type}}, \code{\link{ph_location}}
}
\concept{functions for placeholder location}
