% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_ph_manipulate.R
\name{ph_remove}
\alias{ph_remove}
\title{remove shape}
\usage{
ph_remove(x, type = NULL, id_chr = NULL)
}
\arguments{
\item{x}{a pptx device}

\item{type}{placeholder type}

\item{id_chr}{placeholder id (a string). This is to be used when a placeholder type
is not unique in the current slide, e.g. two placeholders with type 'body'.
Values can be read from \code{\link{slide_summary}}.}
}
\description{
remove a shape in a slide
}
\examples{
fileout <- tempfile(fileext = ".pptx")
doc <- read_pptx()
doc <- add_slide(doc, layout = "Title and Content", master = "Office Theme")
doc <- ph_with_text(x = doc, type = "title", str = "Un titre")
slide_summary(doc) # read column id here
doc <- ph_remove(x = doc, type = "title", id_chr = "2")

if( has_zip() )
  print(doc, target = fileout )
}
