% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{ph_with_gg_at}
\alias{ph_with_gg_at}
\title{add ggplot to a pptx presentation}
\usage{
ph_with_gg_at(x, value, width, height, left, top, ...)
}
\arguments{
\item{x}{an pptx object}

\item{value}{ggplot object}

\item{width, height}{image size in inches}

\item{left, top}{location of the new shape on the slide}

\item{...}{Arguments to be passed to png function.}
}
\description{
add a ggplot as a png image into an rpptx object
This function is deprecated in favor of \code{\link{ph_with}}.
}
