% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_ph_with_methods.R
\name{ph_with_img}
\alias{ph_with_img}
\alias{ph_with_img_at}
\title{add image}
\usage{
ph_with_img(x, src, type = "body", index = 1, width = NULL,
  height = NULL, location = NULL)

ph_with_img_at(x, src, left, top, width, height, rot = 0)
}
\arguments{
\item{x}{an pptx object}

\item{src}{image filename, the basename of the file must not contain any blank.}

\item{type}{placeholder type (i.e. 'body', 'title')}

\item{index}{placeholder index (integer). This is to be used when a placeholder type
is not unique in the current slide, e.g. two placeholders with type 'body', the first
one will be added with index 1 and the second one with index 2.
It is recommanded to use argument \code{location} instead of \code{type} and
\code{index}.}

\item{width, height}{image size in inches}

\item{location}{a placeholder location object. This is a convenient
argument that can replace usage of arguments \code{type} and \code{index}.
See \link{ph_location_type}, \link{ph_location}, \link{ph_location_label},
\link{ph_location_left}, \link{ph_location_right}, \link{ph_location_fullsize}.}

\item{left, top}{location of the new shape on the slide}

\item{rot}{rotation angle}
}
\description{
add an image as a new shape in the current slide.
This function will be deprecated in favor of \code{\link{ph_with}}
in the next release.
}
