% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_ph_with.R, R/ppt_ph_with_at.R
\name{ph_with_img}
\alias{ph_with_img}
\alias{ph_with_img_at}
\title{add image}
\usage{
ph_with_img(x, src, type = "body", index = 1, width = NULL,
  height = NULL)

ph_with_img_at(x, src, left, top, width, height, rot = 0)
}
\arguments{
\item{x}{a pptx device}

\item{src}{image path}

\item{type}{placeholder type}

\item{index}{placeholder index (integer). This is to be used when a placeholder type
is not unique in the current slide, e.g. two placeholders with type 'body'.}

\item{width, height}{image size in inches}

\item{left, top}{location of the new shape on the slide}

\item{rot}{rotation angle}
}
\description{
add an image as a new shape in the current slide.
}
\examples{
fileout <- tempfile(fileext = ".pptx")
doc <- read_pptx()
doc <- add_slide(doc, layout = "Title and Content", master = "Office Theme")

img.file <- file.path( Sys.getenv("R_HOME"), "doc", "html", "logo.jpg" )
if( file.exists(img.file) ){
  doc <- ph_with_img(x = doc, type = "body", src = img.file, height = 1.06, width = 1.39 )
}

print(doc, target = fileout )

fileout <- tempfile(fileext = ".pptx")
doc <- read_pptx()
doc <- add_slide(doc, layout = "Title and Content", master = "Office Theme")

img.file <- file.path( Sys.getenv("R_HOME"), "doc", "html", "logo.jpg" )
if( file.exists(img.file) ){
  doc <- ph_with_img_at(x = doc, src = img.file, height = 1.06, width = 1.39,
    left = 4, top = 4, rot = 45 )
}

print(doc, target = fileout )
}
