% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_ph_with_methods.R
\name{ph_with_ul}
\alias{ph_with_ul}
\title{add unordered list to a pptx presentation}
\usage{
ph_with_ul(x, type = "body", index = 1, str_list = character(0),
  level_list = integer(0), style = NULL, location = NULL)
}
\arguments{
\item{x}{an pptx object}

\item{type}{placeholder type (i.e. 'body', 'title')}

\item{index}{placeholder index (integer). This is to be used when a placeholder type
is not unique in the current slide, e.g. two placeholders with type 'body', the first
one will be added with index 1 and the second one with index 2.
It is recommanded to use argument \code{location} instead of \code{type} and
\code{index}.}

\item{str_list}{list of strings to be included in the object}

\item{level_list}{list of levels for hierarchy structure}

\item{style}{text style, a \code{fp_text} object list or a
single \code{fp_text} objects. Use \code{fp_text(font.size = 0, ...)} to
inherit from default sizes of the presentation.}

\item{location}{a placeholder location object. This is a convenient
argument that can replace usage of arguments \code{type} and \code{index}.
See \link{ph_location_type}, \link{ph_location}, \link{ph_location_label},
\link{ph_location_left}, \link{ph_location_right}, \link{ph_location_fullsize}.}
}
\description{
add an unordered list of text
into an rpptx object. Each text is associated with
a hierarchy level.
This function will be deprecated in favor of \code{\link{ph_with}}
in the next release.
}
