% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_ph_annotate.R
\name{phs_annotate}
\alias{phs_annotate}
\title{Annotate a slide's phs with type, label, and id}
\usage{
phs_annotate(
  x,
  ...,
  .slide_idx = NULL,
  .font_size = NA,
  .font_color = NA,
  .bg = NA,
  .line = sp_line(lwd = 1, color = "grey", lty = "dash"),
  .keys = TRUE,
  .z_position = "front",
  .inform = TRUE
)
}
\arguments{
\item{x}{A \code{rpptx} object.}

\item{...}{Placeholders to annotate. Either short form location syntax (see \code{\link[=phs_with]{phs_with()}} for details) or a
\code{\link[=ph_location]{ph_location()}} object.}

\item{.slide_idx}{Indexes of slides to process. Default (\code{NULL}) is the current slide. Use \code{"all"} to process all
slides.}

\item{.font_size}{Named vector with font sizes for \code{labels}, \code{type}, and \code{id}. A single numeric
value will apply to all three parameters. Matching by position and partial name matching is supported. \code{NA} will
use the  ph's default font size.}

\item{.font_color}{Named vector of font colors for \code{labels}, \code{type}, and \code{id}. Default is \code{c(label = "red", type = "blue", id = "darkgreen")}.  A single numeric value will apply to all three parameters. Matching by position and
partial name matching is supported. \code{NA} will use the  ph's default color.}

\item{.bg}{Background color as hex value or valid R color.}

\item{.line}{Line around placeholder. Either a \code{\link[=sp_line]{sp_line()}} or a single color value (hex or valid R color).}

\item{.keys}{Add keys to info (\verb{id=, type=, label=}) (default is \code{TRUE}).}

\item{.z_position}{Place boxes in the \code{"back"} (default) or \code{"front"}. Only relevant if ph has already been filled.}

\item{.inform}{Inform if a placeholder from \code{...} was not found.}
}
\description{
Draws a box with ph info (type, label, and id) at one or more placeholder position. While \code{\link[=plot_layout_properties]{plot_layout_properties()}}
generates an image of the slide layout, \code{\link[=phs_annotate]{phs_annotate()}} modifies the presentation object.
}
\details{
Throws a warning if one or more phs cannot be found.
}
\examples{
# Annotate all placeholders on current slide
x <- read_pptx()
x <- add_slide(x, "Title Slide")
x <- phs_annotate(x)
\dontrun{
print(x, preview = TRUE)}


# Annotate all placeholders on all slides
x <- read_pptx()
x <- add_slide(x, "Title and Content")
x <- add_slide(x, "Two Content")
x <- phs_annotate(x, .slide_idx = "all") # annotate all slides
\dontrun{
print(x, preview = TRUE)}


# Annotate some placeholders only
x <- read_pptx()
x <- add_slide(x, "Title Slide")
x <- phs_annotate(x, "Title 1") # annotate phs with label "Title 1" only
x <- add_slide(x, "Title and Content")
x <- add_slide(x, "Two Content")
x <- phs_annotate(x, "dt", "ftr", .slide_idx = 2:3) # only types "dt" and "ftr" on slides 2 and 3
\dontrun{
print(x, preview = TRUE)}


# Modify ph appearance (usually not needed)
x <- read_pptx()
x <- add_slide(x, "Title Slide")
x <- phs_annotate(x, .font_color = c("red", "blue", "darkgreen"), .font_size = 12,
                  .bg = "#ff000010", .line = NA)
\dontrun{
print(x, preview = TRUE)}
}
\seealso{
\code{\link[=plot_layout_properties]{plot_layout_properties()}}

Other functions for reading presentation information:
\code{\link{add_annotated_layouts}()},
\code{\link{annotate_base}()},
\code{\link{color_scheme}()},
\code{\link{doc_properties}()},
\code{\link{layout_properties}()},
\code{\link{layout_summary}()},
\code{\link{length.rpptx}()},
\code{\link{plot_layout_properties}()},
\code{\link{slide_size}()},
\code{\link{slide_summary}()}
}
\concept{ppt_info}
