% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_ph_annotate.R
\name{phs_annotate}
\alias{phs_annotate}
\title{Annotate a slide's phs with type, label, and id}
\usage{
phs_annotate(
  x,
  ...,
  .slide_idx = NULL,
  .font_size = 12,
  .font_color = c(label = "red", type = "blue", id = "darkgreen"),
  .bg = "#0000FF10",
  .line = sp_line(lwd = 1, color = "blue", lty = "dash"),
  .keys = TRUE,
  .z_position = "back",
  .inform = TRUE
)
}
\arguments{
\item{x}{A \code{rpptx} object.}

\item{...}{Placeholders to annotate. Either short form location syntax (see \code{\link[=phs_with]{phs_with()}} for details) or a
\code{\link[=ph_location]{ph_location()}} object.}

\item{.slide_idx}{Indexes of slides to process. Default (\code{NULL}) is the current slide. \code{"all"} will process all
slides.}

\item{.font_size}{Named vector with font sizes for \code{labels}, \code{type}, and \code{id}. Default is \code{12}. A single numeric
value will apply to all three parameters. Matching by position and partial name matching is supported. \code{NA} will
use the  ph's default font size.}

\item{.font_color}{Named vector of font colors for \code{labels}, \code{type}, and \code{id}. Default is \code{c(label = "red", type = "blue", id = "darkgreen")}.  A single numeric value will apply to all three parameters. Matching by position and
partial name matching is supported. \code{NA} will use the  ph's default color.}

\item{.bg}{Background color as hex value or valid R color.}

\item{.line}{Line around placeholder. Either a \code{\link[=sp_line]{sp_line()}} or a single color value (hex or valid R color).}

\item{.keys}{Add keys to info (\verb{id=, type=, label=}) (default is \code{TRUE}).}

\item{.z_position}{Place boxes in the \code{"back"} (default) or \code{"front"}.}

\item{.inform}{Inform if a placeholder in \link{...} was not found.}
}
\description{
Draws a box with ph info (type, label, and id) at one or more placeholder position. While \code{\link[=plot_layout_properties]{plot_layout_properties()}}
generates an image of the slide layout, \code{\link[=phs_annotate]{phs_annotate()}} modifies the presentation object.
}
\details{
Throws a warning if one or more phs cannot be found.
}
\examples{
file_pptx <- tempfile(fileext = ".pptx")

x <- read_pptx()

x <- add_slide(x, "Title Slide", ctrTitle = "My Title")
x <- phs_annotate(x) # annotate all phs on current slide

x <- add_slide(x, "Title and Content", body = plot_instr(plot(0:10)))
x <- phs_annotate(x, "dt", "Title 1") # annotate phs with tyope dt and label "Title 1

x <- add_slide(x, "Title and Content")
x <- add_slide(x, "Title and Content")
x <- phs_annotate(x, "sldNum", "body", .slide_idx = 3:4, .font_color = NA) # default color and size

print(x, file_pptx)
# browseUR(file_pptx)  # may not work on all systems

x <- annotate_layouts()
print(x, file_pptx)
# browseURL(file_pptx)  # may not work on all systems
}
\seealso{
\code{\link[=plot_layout_properties]{plot_layout_properties()}}
}
