% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_placeholder.R
\name{placeholder_add_paragraph}
\alias{placeholder_add_paragraph}
\title{append text}
\usage{
placeholder_add_paragraph(x, type = NULL, id_chr = NULL, level = 1)
}
\arguments{
\item{x}{a pptx device}

\item{type}{placeholder type}

\item{id_chr}{placeholder id (a string). This is to be used when a placeholder type
is not unique in the current slide, e.g. two placeholders with type 'body'.
Values can be read from \code{\link{slide_summary}}.}

\item{level}{paragraph level}
}
\description{
append text into a paragraph of a pptx object
}
\examples{
library(magrittr)

fileout <- tempfile(fileext = ".pptx")
default_text <- fp_text(font.size = 0, bold = TRUE, color = "red")

doc <- read_pptx() \%>\%
  add_slide(layout = "Title and Content", master = "Office Theme") \%>\%
  placeholder_set_text(type = "body", str = "A text") \%>\%
  placeholder_add_paragraph(level = 2) \%>\%
  placeholder_add_text(str = "and another, ", style = default_text ) \%>\%
  placeholder_add_paragraph(level = 3) \%>\%
  placeholder_add_text(str = "and another!",
    style = update(default_text, color = "blue")) \%>\%
  print(target = fileout)
}

