% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_placeholder.R
\name{placeholder_add_text}
\alias{placeholder_add_text}
\title{append text}
\usage{
placeholder_add_text(x, str, uid, style = fp_text(font.size = 0),
  pos = "after")
}
\arguments{
\item{x}{a pptx device}

\item{str}{text to add}

\item{uid}{unique id, use column \code{id} of the
result of \code{\link{summary.pptx}}}

\item{style}{text style, a \code{\link{fp_text}} object}

\item{pos}{where to add the new element relative to the cursor,
"after" or "before".}
}
\description{
append text into a paragraph of a pptx object
}
\examples{
library(magrittr)
fileout <- tempfile(fileext = ".pptx")
doc <- pptx() \%>\%
  add_slide(layout = "Title and Content", master = "Office Theme") \%>\%
  placeholder_set_text(id = "body", str = "Un texte")
id_body <- summary(doc)$id[1]
placeholder_add_text(doc, str = " et un autre ", uid = id_body) \%>\%
  print(target = fileout)
}

