% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx_informations.R
\name{plot_layout_properties}
\alias{plot_layout_properties}
\title{Slide layout properties plot}
\usage{
plot_layout_properties(
  x,
  layout = NULL,
  master = NULL,
  labels = TRUE,
  title = TRUE,
  type = TRUE,
  id = TRUE,
  cex = NULL
)
}
\arguments{
\item{x}{an \code{rpptx} object}

\item{layout}{slide layout name.}

\item{master}{master layout name where \code{layout} is located.}

\item{labels}{if \code{TRUE} (default), placeholder labels will be printed (in \emph{red}).}

\item{title}{if \code{TRUE} (default), a title with the layout name will be printed at the top.}

\item{type}{if \code{TRUE} (default), the placeholder type and its index (in square brackets)
is printed in the upper left corner (in \emph{blue}).}

\item{id}{if \code{TRUE} (default), the placeholder's unique \code{id} (see column \code{id} from
\link{layout_properties}) is printed in the upper right corner in \emph{green}.}

\item{cex}{Named list or vector to specify font size for \code{labels}, \code{type}, and \code{id}. Default is
\code{c(labels = .5, type = .5, id = .5)}.}
}
\description{
Plot slide layout properties into corresponding placeholders.
This can be useful to help visualize placeholders locations and identifiers.
\emph{All} information in the plot stems from the \link{layout_properties} output. See \emph{details section}
for more info.
}
\details{
The plot contains all relevant information to address a placeholder using the \verb{ph_location_*}
function family.
\itemize{
\item \code{label}: ph label (red, center) to be used in \link{ph_location_label}.
\item \code{type[idx]}: ph type + type index in brackets (blue, upper left) to be used in \link{ph_location_type}.
\item \code{id}: ph id (green, upper right) to be used in \code{ph_location_id} (forthcoming).
}
}
\examples{
x <- read_pptx()
plot_layout_properties(x = x, layout = "Title Slide", master = "Office Theme")
plot_layout_properties(x = x, layout = "Two Content")
plot_layout_properties(x = x, layout = "Two Content", title = FALSE, type = FALSE, id = FALSE)
plot_layout_properties(x = x, layout = "Two Content", cex = c(labels = 1, id = .7, type = .7))

}
\seealso{
Other functions for reading presentation information: 
\code{\link{annotate_base}()},
\code{\link{color_scheme}()},
\code{\link{doc_properties}()},
\code{\link{layout_properties}()},
\code{\link{layout_summary}()},
\code{\link{length.rpptx}()},
\code{\link{slide_size}()},
\code{\link{slide_summary}()}
}
\concept{functions for reading presentation information}
