% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_pptx.R
\name{pptx_summary}
\alias{pptx_summary}
\title{get PowerPoint content in a data.frame}
\usage{
pptx_summary(x)
}
\arguments{
\item{x}{an rpptx object}
}
\description{
read content of a PowerPoint document and
return a dataset representing the document.
}
\examples{
example_pptx <- system.file(package = "officer",
  "doc_examples/example.pptx")
doc <- read_pptx(example_pptx)
pptx_summary(doc)
pptx_summary(example_pptx)
}
