% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pptx.R
\name{print.rpptx}
\alias{print.rpptx}
\title{write a 'PowerPoint' file.}
\usage{
\method{print}{rpptx}(x, target = NULL, ...)
}
\arguments{
\item{x}{an rpptx object}

\item{target}{path to the pptx file to write}

\item{...}{unused}
}
\description{
write a 'PowerPoint' file.
}
\examples{
# write a rdocx object in a docx file ----
file <- tempfile(fileext = ".pptx")
doc <- read_pptx()
print(doc, target = file)
}
\seealso{
\code{\link{read_pptx}}
}
