% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{read_data}
\alias{read_data}
\alias{read_data.default}
\title{Read common data formats}
\usage{
read_data(file, ...)

\method{read_data}{default}(file, ..., sheet = NULL, delim = NULL,
  encoding = "UTF-8")
}
\arguments{
\item{file}{Path to a file with a supported extension.}

\item{...}{Additional arguments passed to underlying functions.}

\item{sheet}{Specify one or more sheets when reading excel files.}

\item{delim}{The deliminator used for flat files.}

\item{encoding}{The current encoding for the file. Defaults to \code{UTF-8}.}
}
\value{
A data.frame. If more than one sheet is read from a xlsx file
(or you are reading a Rdata file) a list is returned instead.
}
\description{
A simple wrapper for reading data which currently supports Rdata, sav, txt,
csv, csv2 and xlsx. Under the hood, it uses \pkg{readxl}, \pkg{readr} and
\pkg{haven}.
}
\examples{
\dontrun{
df <- data.frame("String" = c("A", "B"), "Int" = c(1:2L), "Percent" = c(0.5, 0.75))
write_data(df, file = "example data.xlsx")
read_data(df, file = "example data.xlsx")
}
}
\author{
Kristian D. Olsen
}

