% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_run_objects.R
\name{run_autonum}
\alias{run_autonum}
\title{auto number}
\usage{
run_autonum(seq_id = "table", pre_label = "TABLE ", post_label = ": ")
}
\arguments{
\item{seq_id}{sequence identifier}

\item{pre_label, post_label}{text to add before and after number}
}
\description{
Create a string representation of a number
}
\examples{
run_autonum()
run_autonum(seq_id = "fig", pre_label = "fig. ")
}
