% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_run_objects.R
\name{run_reference}
\alias{run_reference}
\title{reference}
\usage{
run_reference(id)
}
\arguments{
\item{id}{reference id, a string}
}
\description{
Create a representation of a reference
}
\examples{
run_reference('a_ref')
}
\seealso{
Other run functions for reporting: 
\code{\link{external_img}()},
\code{\link{ftext}()},
\code{\link{run_autonum}()},
\code{\link{run_columnbreak}()},
\code{\link{run_linebreak}()},
\code{\link{run_pagebreak}()},
\code{\link{run_seqfield}()}
}
\concept{run functions for reporting}
