% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_run_objects.R
\name{run_reference}
\alias{run_reference}
\title{reference}
\usage{
run_reference(id, prop = NULL)
}
\arguments{
\item{id}{reference id, a string}

\item{prop}{formatting text properties returned by \link{fp_text}.}
}
\description{
Create a representation of a reference
}
\section{usage}{

You can use this function in conjunction with \link{fpar} to create paragraphs
consisting of differently formatted text parts. You can also use this
function as an \emph{r chunk} in an R Markdown document made with package
officedown.
}

\examples{
run_reference('a_ref')
}
\seealso{
Other run functions for reporting: 
\code{\link{external_img}()},
\code{\link{ftext}()},
\code{\link{hyperlink_ftext}()},
\code{\link{run_autonum}()},
\code{\link{run_bookmark}()},
\code{\link{run_columnbreak}()},
\code{\link{run_footnoteref}()},
\code{\link{run_footnote}()},
\code{\link{run_linebreak}()},
\code{\link{run_pagebreak}()},
\code{\link{run_word_field}()}

Other Word computed fields: 
\code{\link{run_autonum}()},
\code{\link{run_word_field}()}
}
\concept{Word computed fields}
\concept{run functions for reporting}
