% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_run_objects.R
\name{run_seqfield}
\alias{run_seqfield}
\title{seqfield}
\usage{
run_seqfield(seqfield)
}
\arguments{
\item{seqfield}{seqfield string}
}
\description{
Create a seqfield
}
\seealso{
Other run functions for reporting: 
\code{\link{external_img}()},
\code{\link{ftext}()},
\code{\link{run_autonum}()},
\code{\link{run_columnbreak}()},
\code{\link{run_linebreak}()},
\code{\link{run_pagebreak}()},
\code{\link{run_reference}()}
}
\concept{run functions for reporting}
