% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_run_objects.R
\name{run_word_field}
\alias{run_word_field}
\alias{run_seqfield}
\title{seqfield}
\usage{
run_word_field(field, prop = NULL, seqfield = field)

run_seqfield(field, prop = NULL, seqfield = field)
}
\arguments{
\item{field, seqfield}{computed field string (\code{seqfield} will be
totally superseded by \code{field} in the futur).}

\item{prop}{formatting text properties returned by \link{fp_text}.}
}
\description{
Create a Word computed field.
}
\note{
In the previous version, this function was called \code{run_seqfield}
but the name was wrong and should have been \code{run_word_field}.
}
\seealso{
Other run functions for reporting: 
\code{\link{external_img}()},
\code{\link{ftext}()},
\code{\link{run_autonum}()},
\code{\link{run_columnbreak}()},
\code{\link{run_linebreak}()},
\code{\link{run_pagebreak}()},
\code{\link{run_reference}()}

Other Word computed fields: 
\code{\link{run_autonum}()},
\code{\link{run_reference}()}
}
\concept{Word computed fields}
\concept{run functions for reporting}
