% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_notes.R
\name{set_notes}
\alias{set_notes}
\alias{set_notes.character}
\title{set notes for current slide slide}
\usage{
set_notes(x, value, location, ...)

\method{set_notes}{character}(x, value, location, ...)
}
\arguments{
\item{x}{an rpptx object}

\item{value}{text to be added to notes}

\item{location}{a placeholder location object.
It will be used to specify the location of the new shape. This location
can be defined with a call to one of the notes_ph functions. See
section \code{"see also"}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
set speaker notes for the current slide in a pptx presentation
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: add a character vector to a place holder in the notes on the
current slide, values will be added as paragraphs.
}}

\examples{
# this name will be used to print the file
# change it to "youfile.pptx" to write the pptx
# file in your working directory.
fileout <- tempfile(fileext = ".pptx")

doc <- read_pptx()
# add a slide with some text ----
doc <- add_slide(doc, layout = "Title and Content", master = "Office Theme")
doc <- ph_with(x = doc, value = "Slide Title",
   location = ph_location_type(type = "title") )
# set speaker notes for the slide ----
doc <- set_notes(doc, value = "This text will only be visible for the speaker.",
   location = notes_location_type("body"))

print(doc, target = fileout)

}
\seealso{
\code{\link[=print.rpptx]{print.rpptx()}}, \code{\link[=read_pptx]{read_pptx()}}, \code{\link[=add_slide]{add_slide()}}, \code{\link[=notes_location_label]{notes_location_label()}}, \code{\link[=notes_location_type]{notes_location_type()}}

Other functions slide manipulation: 
\code{\link{add_slide}()},
\code{\link{move_slide}()},
\code{\link{on_slide}()},
\code{\link{remove_slide}()}
}
\concept{functions slide manipulation}
