% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx_slide_manip.R
\name{shape_properties_tags}
\alias{shape_properties_tags}
\title{pptx tags for visual and non visual properties}
\usage{
shape_properties_tags(
  left = 0,
  top = 0,
  width = 3,
  height = 3,
  bg = "transparent",
  rot = 0,
  label = "",
  ph = "<p:ph/>",
  ln = sp_line(lwd = 0, linecmpd = "solid", lineend = "rnd"),
  geom = NULL
)
}
\arguments{
\item{left, top, width, height}{place holder coordinates
in inches.}

\item{bg}{background color}

\item{rot}{rotation angle}

\item{label}{a label for the placeholder.}

\item{ph}{string containing xml code for ph tag}

\item{ln}{a \code{\link[=sp_line]{sp_line()}} object specifying the outline style.}

\item{geom}{shape geometry, see http://www.datypic.com/sc/ooxml/t-a_ST_ShapeType.html}
}
\value{
a character value
}
\description{
Visual and non visual properties of a
shape can be returned by this function.
}
\seealso{
Other functions for officer extensions: 
\code{\link{fortify_location}()},
\code{\link{get_reference_value}()},
\code{\link{opts_current_table}()},
\code{\link{str_encode_to_rtf}()},
\code{\link{to_html}()},
\code{\link{to_pml}()},
\code{\link{to_rtf}()},
\code{\link{to_wml}()},
\code{\link{wml_link_images}()}
}
\concept{functions for officer extensions}
\keyword{internal}
