% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xlsx.R
\name{sheet_add_drawing}
\alias{sheet_add_drawing}
\title{Add a drawing to an Excel sheet}
\usage{
sheet_add_drawing(x, value, sheet, ...)
}
\arguments{
\item{x}{rxlsx object created by \code{\link[=read_xlsx]{read_xlsx()}}}

\item{value}{object to add (dispatched to the appropriate method)}

\item{sheet}{sheet name (must already exist, see \code{\link[=add_sheet]{add_sheet()}})}

\item{...}{additional arguments passed to methods}
}
\value{
the rxlsx object (invisibly)
}
\description{
Add a graphical element into a sheet of an xlsx workbook.
This is a generic function dispatching on \code{value}. Methods are
provided by extension packages 'mschart' and 'rvg'.

Use \code{\link[=sheet_write_data]{sheet_write_data()}} to write data into the sheet before or
after adding a drawing.
}
\seealso{
\code{\link[=read_xlsx]{read_xlsx()}}, \code{\link[=add_sheet]{add_sheet()}}, \code{\link[=sheet_write_data]{sheet_write_data()}}
}
