% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorcuts.R
\name{fp_bold}
\alias{fp_bold}
\alias{shortcut_properties}
\alias{fp_italic}
\alias{b_null}
\title{shortcuts for formatting properties}
\usage{
fp_bold(...)

fp_italic(...)

b_null(...)
}
\arguments{
\item{...}{further arguments passed to original functions.}
}
\description{
Shortcuts for \code{fp_text}, \code{fp_par},
\code{fp_cell} and \code{fp_border}.
}
\examples{
fp_bold()
fp_italic()
b_null()
}

