% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorcuts.R
\name{t_b}
\alias{b_dash}
\alias{b_dot}
\alias{b_null}
\alias{c_b_all}
\alias{c_b_b}
\alias{c_b_lr}
\alias{c_b_null}
\alias{c_b_t}
\alias{c_b_tb}
\alias{p_c}
\alias{p_j}
\alias{p_l}
\alias{p_r}
\alias{shortcut_properties}
\alias{t_b}
\alias{t_bi}
\alias{t_i}
\title{shortcuts for formatting properties}
\usage{
t_b(...)

t_i(...)

t_bi(...)

p_r(...)

p_l(...)

p_c(...)

p_j(...)

b_dot(...)

b_dash(...)

b_null(...)

c_b_null(...)

c_b_all(...)

c_b_b(...)

c_b_t(...)

c_b_tb(...)

c_b_lr(...)
}
\arguments{
\item{...}{further arguments passed to original functions.}
}
\description{
Shortcuts for \code{pr_text}, \code{pr_par},
\code{pr_cell} and \code{pr_border}.
}
\examples{
t_b()
t_i()
t_bi()
p_r()
p_l()
p_c()
p_j()
b_dot()
b_dash()
b_null()
c_b_null()
c_b_all()
c_b_b()
c_b_t()
c_b_tb()
c_b_tb()
}

