% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorcuts.R
\name{shortcuts}
\alias{shortcuts}
\title{shortcuts for formatting properties}
\usage{
shortcuts
}
\description{
Shortcuts for \code{fp_text}, \code{fp_par},
\code{fp_cell} and \code{fp_border}.
}
\examples{
shortcuts$fp_bold()
shortcuts$fp_italic()
shortcuts$b_null()
}
