% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pptx.R
\name{slide_summary}
\alias{slide_summary}
\title{slide summary}
\usage{
slide_summary(x, index = NULL)
}
\arguments{
\item{x}{pptx object}

\item{index}{slide index}
}
\description{
get informations about current slide
into a data.frame.
}
\examples{
library(magrittr)

my_pres <- read_pptx() \%>\%
  add_slide(layout = "Two Content", master = "Office Theme") \%>\%
  ph_with_text(type = "dt", str = format(Sys.Date())) \%>\%
  add_slide(layout = "Title and Content", master = "Office Theme")

slide_summary(my_pres)
slide_summary(my_pres, index = 1)
}
