% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{slip_in_column_break}
\alias{slip_in_column_break}
\title{add a column break}
\usage{
slip_in_column_break(x, pos = "before")
}
\arguments{
\item{x}{an rdocx object}

\item{pos}{where to add the new element relative to the cursor,
"after" or "before".}
}
\description{
add a column break into a Word document. A column break
is used to add a break in a multi columns section in a Word
Document.

This function will be deprecated in the next release because it is not
efficient and make users write complex code, use \code{\link[=run_columnbreak]{run_columnbreak()}} instead.
}
