% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{slip_in_footnote}
\alias{slip_in_footnote}
\title{append a footnote}
\usage{
slip_in_footnote(x, ...)
}
\arguments{
\item{x}{an rdocx object}

\item{...}{unused}
}
\description{
this function is deprecated.
}
\note{
This function is now deprecated because it is not
efficient and make users write complex code. Use instead \code{\link[=fpar]{fpar()}} to build
formatted paragraphs.
}
\keyword{internal}
