% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{slip_in_text}
\alias{slip_in_text}
\title{append text}
\usage{
slip_in_text(x, str, style = NULL, pos = "after", hyperlink = NULL)
}
\arguments{
\item{x}{an rdocx object}

\item{str}{text}

\item{style}{text style}

\item{pos}{where to add the new element relative to the cursor,
"after" or "before".}

\item{hyperlink}{turn the text into an external hyperlink}
}
\description{
append text into a paragraph of an rdocx object.

This function will be deprecated in the next release because it is not
efficient and make users write complex code. Use instead \code{\link[=fpar]{fpar()}} to build
formatted paragraphs.
}
