% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cot_objects.R
\name{stext}
\alias{stext}
\title{Word styled text}
\usage{
stext(text, style_id = NULL, style, doc)
}
\arguments{
\item{text}{text value}

\item{style_id}{style id}

\item{style}{style name. Use only when you don't know what is the id
associated with a stylename.}

\item{doc}{\code{rdocx} object where the text will be inserted.
This is only used to get the style id from the style name (use
only when you don't know what is the id associated with a stylename).}
}
\description{
Format a chunk of text with a referenced style.
}
\examples{
doc <- read_docx()
stext("hello", "strong", doc)
}
