% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_docx.R
\name{styles_info}
\alias{styles_info}
\title{read Word styles}
\usage{
styles_info(x)
}
\arguments{
\item{x}{an rdocx object}
}
\description{
read Word styles and get results in
a tidy data.frame.
}
\examples{
x <- read_docx()
styles_info(x)
}
\seealso{
Other functions for Word document informations: 
\code{\link{doc_properties}()},
\code{\link{docx_bookmarks}()},
\code{\link{docx_dim}()},
\code{\link{length.rdocx}()},
\code{\link{set_doc_properties}()}
}
\concept{functions for Word document informations}
