% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_docx.R
\name{styles_info}
\alias{styles_info}
\title{read Word styles}
\usage{
styles_info(x)
}
\arguments{
\item{x}{a rdocx object}
}
\description{
read Word styles and get results in
a tidy data.frame.
}
\examples{
library(magrittr)
read_docx() \%>\% styles_info()
}
