% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_block_objects.R
\name{table_colwidths}
\alias{table_colwidths}
\title{Column widths of a table}
\usage{
table_colwidths(widths = NULL)
}
\arguments{
\item{widths}{Column widths expressed in inches.}
}
\description{
The function defines the size of each column of a table.
}
\seealso{
Other functions for table definition: 
\code{\link{prop_table}()},
\code{\link{table_conditional_formatting}()},
\code{\link{table_layout}()},
\code{\link{table_width}()}
}
\concept{functions for table definition}
