% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_block_objects.R
\name{table_width}
\alias{table_width}
\title{Preferred width for a table}
\usage{
table_width(width = 1, unit = "pct")
}
\arguments{
\item{width}{value of the preferred width of the table.}

\item{unit}{unit of the width. Possible values are 'in' (inches) and 'pct' (percent)}
}
\description{
Define the preferred width for a table.
}
\section{Word}{

All widths in a table are considered preferred because widths of
columns can conflict and the table layout rules can require a
preference to be overridden.
}

\seealso{
Other functions for table definition: 
\code{\link{prop_table}()},
\code{\link{table_colwidths}()},
\code{\link{table_conditional_formatting}()},
\code{\link{table_layout}()},
\code{\link{table_stylenames}()}
}
\concept{functions for table definition}
