% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_run_objects.R
\name{to_html}
\alias{to_html}
\title{Convert officer objects to HTML}
\usage{
to_html(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{Arguments to be passed to methods}
}
\description{
Convert an object made with package officer
to HTML. The result is a string.
}
