% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled.R
\name{to_labelled}
\alias{to_labelled}
\alias{to_labelled.data.frame}
\alias{to_labelled.data.table}
\title{Convert to labelled}
\usage{
to_labelled(x)

\method{to_labelled}{data.frame}(x)

\method{to_labelled}{data.table}(x)
}
\arguments{
\item{x}{A data.frame, or \code{\link[reporttoolDT]{Survey}}.}
}
\description{
Reverses the process from \code{\link{from_labelled}}, by attempting to create
labelled variables in place of \code{\link[base]{factor}}, and adding labels to each variable.
}
\note{
Because of a limitation in \pkg{ReadStat} (it can't write strings longer
than 256 characters), \code{\link{write_data}} will write the long strings as
a separate .Rdata file. If you use \code{\link{read_data}}, you will get them back.
}
\examples{
# TODO
}
\author{
Kristian D. Olsen
}

