% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooxml_run_objects.R
\name{to_wml}
\alias{to_wml}
\title{Convert officer objects to WordprocessingML}
\usage{
to_wml(x, add_ns = FALSE, ...)
}
\arguments{
\item{x}{a location for a placeholder.}

\item{add_ns}{should namespace be added to the top tag}

\item{...}{Arguments to be passed to methods}
}
\description{
Convert an object made with package officer
to WordprocessingML. The result is a string.
}
