% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-proc.R
\name{wml_link_images}
\alias{wml_link_images}
\title{transform an xml string with images references}
\usage{
wml_link_images(x, str)
}
\arguments{
\item{x}{an rdocx object}

\item{str}{wml string}
}
\description{
This function is useless now, as the processing of images
is automated when using \code{\link[=print.rdocx]{print.rdocx()}}.
}
\seealso{
Other functions for officer extensions: 
\code{\link{docx_reference_img}()},
\code{\link{fortify_location}()},
\code{\link{get_reference_value}()},
\code{\link{opts_current_table}()},
\code{\link{to_html}()},
\code{\link{to_pml}()},
\code{\link{to_wml}()}
}
\concept{functions for officer extensions}
\keyword{internal}
