% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_img.R
\name{wml_link_images}
\alias{wml_link_images}
\title{transform an xml string with images references}
\usage{
wml_link_images(x, str)
}
\arguments{
\item{x}{an rdocx object}

\item{str}{wml string}
}
\description{
The function replace images filenames
in an xml string with their id. The wml code cannot
be valid without this operation.
}
\details{
The function is available to allow the creation of valid
wml code containing references to images.
}
