% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_output.R
\name{write_data}
\alias{write_data}
\alias{write_data.data.frame}
\alias{write_data.ggplot}
\alias{write_data.list}
\title{Write common file formats}
\usage{
write_data(x, file, ..., delim = NULL)

\method{write_data}{data.frame}(x, file, ..., delim = NULL)

\method{write_data}{list}(x, file, ...)

\method{write_data}{ggplot}(x, file, ...)
}
\arguments{
\item{x}{The data to be written. (\code{data.frame} or \code{list}.
\code{matrix} and \code{table} will be coerced to a \code{data.frame}.}

\item{file}{Path and filename of output.}

\item{...}{Further arguments passed to underlying functions.
(Currently: \code{\link[base]{save}}, \code{\link[openxlsx]{saveWorkbook}},
\code{\link[ggplot2]{ggsave}}, \code{\link[readr]{write_delim}} and \code{\link[haven]{write_sav}}.)}

\item{delim}{The delimiter to use when writing flat files.}
}
\description{
A simple wrapper for writing common data formats. The format is determined
by the extension given in \code{file}. Flat files are written with \pkg{readr},
and the encoding is always \code{UTF-8}. For xlsx, the function uses
\code{\link{to_excel}} (which in turn uses \pkg{openxlsx}).
}
\note{
Use \code{lapply} to write a list of data to flat files (csv, txt etc).
}
\examples{
\dontrun{
df <- data.frame("String" = c("A", "B"), "Int" = c(1:2L), "Percent" = c(0.5, 0.75))
write_data(df, file = "example data.csv")
}
}
\author{
Kristian D. Olsen
}

